### 2020-09-02 #2633 @whedon
[PRE REVIEW]: DeepReg: a deep learning toolkit for medical image registration
**Submitting author:** @mathpluscode (<a href="http://orcid.org/0000-0002-1184-7421">Yunguan Fu</a>)
**Repository:** <a href="https://github.com/DeepRegNet/DeepReg" target ="_blank">https://github.com/DeepRegNet/DeepReg</a>
**Version:** v0.1.0
