### 2020-09-02 #2633 @kthyng
Hi @mathpluscode! Thanks for your submission. There are a few things to take care of before proceeding with the review:

- Need a statement of need in the paper
- your paper is far too long. As stated in the [docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), it should be 250–1000 words.
- you can see the doi's listed above are missing; please fix this.
