### 2020-09-02 #2633 @kthyng
@arokem Do you have bandwidth for handling a JOSS submission right now? We have you marked as "somewhat able" but I seem to remember maybe that is overstating it.
