### 2020-09-02 DeepRegNet/DeepReg#399 @mathpluscode
Fix paper acc to JOSS reviews
## Subject of the documentation

- [x] Need a statement of need in the paper
- [x] your paper is far too long. As stated in the docs, it should be 250–1000 words.
- [x] you can see the doi's listed above are missing; please fix this.

https://github.com/openjournals/joss-reviews/issues/2633
