### 2020-09-02 4e19193
Issue #399: add statement of need
