### 2020-09-02 73843ff
Issue #399: integrate summary and implementation
