### 2020-09-02 02843c5
Issue #399: cut algorithms and conclusion
