+++ docs/joss_paper/paper.md
@@ -71,31 +71,14 @@ Image fusion is a fundamental task in medical image analysis and computer-assist
 
 *DeepReg* is a Python package that implements multiple registration algorithms including unsupervised-, weakly-supervised- and conditional-segmentation algorithms [@de2019deep; @hu2018label; @hu2019conditional]. Predefined _dataset loaders_ for paired-, unpaired- and grouped images are provided, supporting two different data formats. In addition, DeepReg provides command-line tool options that enable a set of basic and advanced functionalities for model training, prediction and image warping. These implementations, together with their documentation, tutorials and demos, simplify workflows for prototyping and developing novel methodology, utilising latest development in this field and accessing quality research advances.
 
-A MICCAI Educational Challenge has utilised the DeepReg code and demos to explore the link between classical algorithms and deep-learning-based methods [@brown2020introduction], while a recently published research paper investigated temporal changes for prostate cancer patients under active surveillance programme, by using longitudinal image registration adapted from the DeepReg code [@yang2020longitudinal].
+An MICCAI Educational Challenge has utilised the DeepReg code and demos to explore the link between classical algorithms and deep-learning-based methods [@brown2020introduction], while a recently published research paper investigated temporal changes for prostate cancer patients under active surveillance programme, by using longitudinal image registration adapted from the DeepReg code [@yang2020longitudinal].
 
 # Statement of need
 
 Currently, popular packages focusing on deep learning methods for medical imaging, such as NiftyNet [@gibson2018niftynet] and MONAI (https://monai.io/), do not support image registration. Other open-source projects implementing specific published algorithms, such as the VoxelMorph [@balakrishnan2019voxelmorph], are seldom tested or designed for general research and education purposes. Therefore an open-sourced project focusing on image registration with deep learning is much needed.
 
-# Algorithms
-In this section we first summarise several standard pairwise image registration network training strategies that are implemented in the DeepReg package. The networks aim to align a pair of moving- and fixed images such that the moving image can be warped or transformed into the fixed image coordinates. The methodologies adopted in these algorithms are building blocks of many other registration tasks, such as group-wise registration and morphological template construction [@dalca2019learning; @siebert2020deep; @luo2020mvmm].
-
-Unsupervised learning was first developed independently at a number of research groups [@de2019deep; @balakrishnan2019voxelmorph]. Image dissimilarity is measured between the fixed and warped moving images, adapted from the classical registration methods [@hill2001medical]. \autoref{fig:unsupervised} shows a schematic illustration of the unsupervised network training. Image dissimilarity measures include sum-of-squared difference in intensity, normalised cross-correlation, mutual information and their variants.
-
-![Registration network with unsupervised loss.\label{fig:unsupervised}](../source/_images/registration-ddf-nn-unsupervised.png)
-
-Weak supervision which utilises segmented corresponding regions in the medical image pairs was first proposed in a multimodal application [@hu2018label; @hu2018weakly]. \autoref{fig:weakly} shows a schematic illustration of the weakly-supervised network training. Label dissimilarity measures including Dice, Jaccard, mean-squared difference, cross-entropy and their multiscale formulations [@hu2018weakly].
-
-![Registration network with weak supervision loss.\label{fig:weakly}](../source/_images/registration-ddf-nn-weakly-supervised.png)
-
-Combining the unsupervised loss and weak supervision has shown superior registration accuracy, compared with that using unsupervised loss alone [@balakrishnan2019voxelmorph]. As a result, the overall loss is the weighted sum of the image dissimilarity, the label dissimilarity and the deformation regularisation, which encourages smoothness of the predicted transformation. \autoref{fig:combined} shows a schematic illustration of the weakly-supervised network training with unsupervised loss.
-
[…]
