### 2020-09-02 #2633 @mathpluscode
Hi @kthyng, for the reference, does every reference have to have a DOI? Is it ok if some references do not have DOI? Thx!
