### 2020-09-02 127c9e0
Issue #399 text updates [ci skip]
