### 2020-09-02 352e35b
Issue #399: correct typos
