+++ docs/joss_paper/paper.md
@@ -69,17 +69,15 @@ bibliography: paper.bib
 # Summary
 Image fusion is a fundamental task in medical image analysis and computer-assisted intervention. Medical image registration, computational algorithms that align different images together [@hill2001medical], has in recent years turned the research attention towards deep learning. Indeed, the representation ability to learn from population data with deep neural networks has opened new possibilities for improving registration generalisability by mitigating difficulties in designing hand-engineered image features and similarity measures for many real-world clinical applications [@haskins2020deep; @fu2020deep]. In addition, its fast inference can substantially accelerate registration execution for time-critical tasks.
 
-*DeepReg* is a Python package that implements multiple registration algorithms and  a set of predefined _dataset loaders_, supporting two different data formats. DeepReg also provides command-line tool options that enable basic and advanced functionalities for model training, prediction and image warping. These implementations, together with their documentation, tutorials and demos, aim to simplify workflows for prototyping and developing novel methodology, utilising latest development and accessing quality research advances.
+*DeepReg* is a Python package that implements multiple registration algorithms and  a set of predefined _dataset loaders_, supporting two different data formats. DeepReg also provides command-line tool options that enable basic and advanced functionalities for model training, prediction and image warping. These implementations, together with their documentation, tutorials and demos, aim to simplify workflows for prototyping and developing novel methodology, utilising latest development and accessing quality research advances. DeepReg is unit tested and a set of customised contributor guidelines are provided to facilitate community contributions.
 
-An MICCAI Educational Challenge has utilised the DeepReg code and demos to explore the link between classical algorithms and deep-learning-based methods [@brown2020introduction], while a recently published research work investigated temporal changes in prostate cancer imaging, by using a longitudinal registration adapted from the DeepReg code [@yang2020longitudinal].
+A submission to the MICCAI Educational Challenge has utilised the DeepReg code and demos to explore the link between classical algorithms and deep-learning-based methods [@brown2020introduction], while a recently published research work investigated temporal changes in prostate cancer imaging, by using a longitudinal registration adapted from the DeepReg code [@yang2020longitudinal].
 
 # Statement of need
 Currently, popular packages focusing on deep learning methods for medical imaging, such as NiftyNet [@gibson2018niftynet] and MONAI (https://monai.io/), do not support image registration. Other open-source projects implementing specific published algorithms, such as the VoxelMorph [@balakrishnan2019voxelmorph], are seldom tested or designed for general research and education purposes. Therefore an open-sourced project focusing on image registration with deep learning is much needed.
 
 # Implementation
-DeepReg implements a framework for unsupervised learning [@de2019deep; @balakrishnan2019voxelmorph], weakly-supervised learning [@hu2018label; @hu2018weakly] and their combinations and variants, e.g. [@hu2019conditional]. Many options are included for major components of these approaches, such as different image- and label dissimilarity functions, transformation models [@ashburner2007fast; @vercauteren2009diffeomorphic; @hill2001medical], deformation regularisation [@rueckert1999nonrigid] and different neural network architectures [@hu2018weakly; @he2016deep; @simonyan2014very; @hu2018weakly]. The provided dataset loaders adopt staged random sampling strategy to ensure unbiased learning from groups, images and labels [@hu2018weakly; @yang2020longitudinal]. These algorithmic components together with the flexible dataset loaders are building blocks of many other registration tasks, such as group-wise registration and morphological template construction [@dalca2019learning; @siebert2020deep; @luo2020mvmm].
-
[…]
