### 2020-09-02 2f586dd
Issue #399: more tweaks of sentences
