+++ docs/joss_paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'DeepReg: a deep-learning toolkit for medical image registration'
+title: 'DeepReg: a deep learning toolkit for medical image registration'
 tags:
   - Python
   - TensorFlow
@@ -69,7 +69,7 @@ bibliography: paper.bib
 # Summary
 Image fusion is a fundamental task in medical image analysis and computer-assisted intervention. Medical image registration, computational algorithms that align different images together [@hill2001medical], has in recent years turned the research attention towards deep learning. Indeed, the representation ability to learn from population data with deep neural networks has opened new possibilities for improving registration generalisability by mitigating difficulties in designing hand-engineered image features and similarity measures for many real-world clinical applications [@haskins2020deep; @fu2020deep]. In addition, its fast inference can substantially accelerate registration execution for time-critical tasks.
 
-*DeepReg* is a Python package that implements multiple registration algorithms and  a set of predefined _dataset loaders_, supporting two different data formats. DeepReg also provides command-line tool options that enable basic and advanced functionalities for model training, prediction and image warping. These implementations, together with their documentation, tutorials and demos, aim to simplify workflows for prototyping and developing novel methodology, utilising latest development and accessing quality research advances. DeepReg is unit tested and a set of customised contributor guidelines are provided to facilitate community contributions.
+*DeepReg* is a Python package that implements multiple registration algorithms and  a set of predefined _dataset loaders_, supporting both labelled- and unlabelled data. DeepReg also provides command-line tool options that enable basic and advanced functionalities for model training, prediction and image warping. These implementations, together with their documentation, tutorials and demos, aim to simplify workflows for prototyping and developing novel methodology, utilising latest development and accessing quality research advances. DeepReg is unit tested and a set of customised contributor guidelines are provided to facilitate community contributions.
 
 A submission to the MICCAI Educational Challenge has utilised the DeepReg code and demos to explore the link between classical algorithms and deep-learning-based methods [@brown2020introduction], while a recently published research work investigated temporal changes in prostate cancer imaging, by using a longitudinal registration adapted from the DeepReg code [@yang2020longitudinal].
 
@@ -83,16 +83,16 @@ DeepReg implements a framework for unsupervised learning [@de2019deep; @balakris
 In addition to the tutorials and documentation, DeepReg provides a collection of demonstrations, _DeepReg Demos_, using open-accessible data with real-world clinical applications.
 
 ## Paired images
-If images are available in pairs, two dataset loaders are currently available for paired images with- and without labels. Many clinical applications for tracking organ motion and other temporal changes require _intra-subject_ _single-modality_ image registration. Registering lung CT images for the same patient, acquired at expiratory and inspiratory phases [@hering_alessa_2020_3835682], is such an example of both unsupervised (without labels) and combined supervision (trained with additional label dissimilarity based on anatomical segmentation). Furthermore, registering prostate MR, acquired before surgery, and intra-operative ultrasound images is an example of weakly-supervised learning for multimodal image registration [@hu2018weakly]. Another DeepReg Demo illustrates MR-to-ultrasound image registration is to track tissue deformation and brain tumour resection during neurosurgery [@xiao2017resect].
[…]
