### 2020-09-02 11b5e03
Issue #399: remove extra space
