### 2020-09-03 bee8392
Issue #399: add missing DOIs
