+++ docs/joss_paper/paper.bib
@@ -61,6 +61,7 @@
   volume={57},
   pages={226--236},
   year={2019},
+  doi={10.1016/j.media.2019.07.006},
   publisher={Elsevier}
 }
 
@@ -108,6 +109,7 @@
   number={1},
   pages={95--113},
   year={2007},
+  doi={10.1016/j.neuroimage.2007.07.007},
   publisher={Elsevier}
 }
 
