### 2020-09-03 #2633 @mathpluscode
> > For the reference, does every reference have to have a DOI? Is it ok if some references do not have DOI?
>
> Any reference that has a doi needs to include it in your references. It looks like there are two more missing to consider above. If they are correct, they need to be included.
>
> > * We've also changed the title from "DeepReg: a deep-learning toolkit for medical image registration" to "DeepReg: a deep learning toolkit for medical image registration", remove the hyphen between deep and learning. Hope it's fine.
>
> I have updated this.

Hi, @kthyng, we've fixed the DOIs.
