### 2020-09-05 9dacbbc
Issue #404: Updated install path from quick start page.
