### 2020-09-10 #2633 @Kevin-Mattheus-Moerman
@stefanklein @mstaring @thewtex @ljod @fedorov @ethanwharris @mamrehn @billchenxi as editor for the [Journal of Open Source Software (JOSS)](https://joss.theoj.org/) I'd like to invite you to review _"DeepReg: a deep learning toolkit for medical image registration"_.

You can comment here to let me know if you are interested. Thanks!

We have a smooth GitHub based [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) which focuses on [the software :computer: ](https://github.com/DeepRegNet/DeepReg) as well as [the very short paper :page_facing_up: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02633/joss.02633/10.21105.joss.02633.pdf).

This here is a "pre-review issue". Once I've attracted sufficient reviewers I'll initiate a review issue where reviewers have a set of check-boxes to guide them through the review process.

Check out [our documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) to learn more about what it is like to review for JOSS, also here is an example review process for a different submission which you can look at:
https://github.com/openjournals/joss-reviews/issues/2520

Thanks!
