### 2020-09-12 a2dcb70
Issue #398: add log_root param to replace hard-coded logs
