+++ test/unit/test_util.py
@@ -19,6 +19,8 @@ from deepreg.util import (
     save_metric_dict,
 )
 
+log_root = "logs"
+
 
 def test_build_dataset():
     """
@@ -27,12 +29,13 @@ def test_build_dataset():
 
     # init arguments
     config_path = "config/unpaired_labeled_ddf.yaml"
+    log_root = "logs"
     log_dir = "test_build_dataset"
     ckpt_path = ""
 
     # load config
     config, log_dir = build_config(
-        config_path=config_path, log_dir=log_dir, ckpt_path=ckpt_path
+        config_path=config_path, log_root=log_root, log_dir=log_dir, ckpt_path=ckpt_path
     )
 
     # build dataset
@@ -72,16 +75,16 @@ def test_build_log_dir():
     """
 
     # use default timestamp based directory
-    log_dir = build_log_dir(log_dir="")
+    log_dir = build_log_dir(log_root=log_root, log_dir="")
     head, tail = os.path.split(log_dir)
-    assert head == "logs"
+    assert head == log_root
     pattern = re.compile("[0-9]{8}-[0-9]{6}")
     assert pattern.match(tail)
 
     # use custom directory
-    log_dir = build_log_dir(log_dir="custom")
+    log_dir = build_log_dir(log_root=log_root, log_dir="custom")
     head, tail = os.path.split(log_dir)
-    assert head == "logs"
+    assert head == log_root
     assert tail == "custom"
 
 
