### 2020-09-12 f2d10f7
Issue #398: refactor test_build_log_dir using pytest.mark.parametrize
