### 2020-09-18 #2633 @stefanklein
Hi, thanks for the invitation to review. I hadn't heard about this journal. The concept seems very interesting! Unfortunately I won't have time to do this review and thoroughly test the code. A postdoc from my group with experience in image registration and deep learning would be interested in doing the review, so please consider inviting her: @lbrea . Best regards, Stefan
