### 2020-09-24 611abfd
Issue #361: model retraining. Script updates are: change of hyperparameters in train and default epoch number in predict
