### 2020-09-25 59e7154
Bump tensorflow from 2.2 to 2.2.1
Bumps [tensorflow](https://github.com/tensorflow/tensorflow) from 2.2 to 2.2.1.
- [Release notes](https://github.com/tensorflow/tensorflow/releases)
- [Changelog](https://github.com/tensorflow/tensorflow/blob/master/RELEASE.md)
- [Commits](https://github.com/tensorflow/tensorflow/compare/v2.2.0...v2.2.1)

Signed-off-by: dependabot[bot] <support@github.com>
