+++ requirements.txt
@@ -6,7 +6,7 @@ pyyaml>=5.3.1
 matplotlib
 argparse
 tqdm
-tensorflow==2.2
+tensorflow==2.2.1
 pytest>=4.6
 pytest-cov
 pytest-dependency
