### 2020-09-28 803fc52
issue #402: remove integration tests
