+++ .travis.yml
@@ -23,22 +23,10 @@ jobs:
         - make clean html
     - stage: "Tests"
       name: "Unit Tests"
-      script:  pytest --cov-report= --cov=deepreg ./test/unit/
+      script: pytest --cov-report= --cov=deepreg ./test/unit/
       after_success: bash <(curl -s https://codecov.io/bash)
     - name: "Command Tests: Unpaired Unlabeled"
       script:
         - deepreg_train -g "" --config_path config/unpaired_labeled_ddf.yaml --log_dir test
         - deepreg_predict -g "" --ckpt_path logs/test/save/weights-epoch2.ckpt --mode test
         - deepreg_warp --image data/test/nifti/unit_test/moving_image.nii.gz --ddf data/test/nifti/unit_test/ddf.nii.gz --out logs/test_warp/out.nii.gz
-    - name: "Output Tests: Paired Unlabeled"
-      script: travis_wait 30 pytest ./test/output/test_paired_unlabeled.py
-    - name: "Output Tests: Paired Labeled"
-      script: travis_wait 30 pytest ./test/output/test_paired_labeled.py
-    - name: "Output Tests: Unpaired Labeled"
-      script: travis_wait 30 pytest ./test/output/test_unpaired_labeled.py
-    - name: "Output Tests: Unpaired Unlabeled"
-      script: travis_wait 30 pytest ./test/output/test_unpaired_unlabeled.py
-    - name: "Output Tests: Grouped Unlabeled"
-      script: travis_wait 30 pytest ./test/output/test_grouped_unlabeled.py
-    - name: "Output Tests: Grouped Labeled"
-      script: travis_wait 30 pytest ./test/output/test_grouped_labeled.py
