### 2020-09-28 25b7af0
Issue #402: merge linting and pre-commit
