### 2020-10-03 9ce7326
Issue #166: add visualisation tool script with CLI
