+++ vis.py
@@ -1,0 +1,173 @@
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib.animation as animation
+import os
+import argparse
+import tensorflow as tf
+
+os.chdir(r'/home/s-sd/Desktop/deepreg_project/DeepReg')
+from deepreg.dataset.loader.nifti_loader import load_nifti_file
+from deepreg.model.layer_util import warp_image_ddf
+
+
+def string_to_list(string):
+    '''
+    converts a comma separated string to a list of strings
+    also removes leading or trailing spaces from each element in list
+    '''
+    return [elem.strip() for elem in string.split(',')]
+
+def gif_slices(img_paths, save_path="", interval=50):
+    if type(img_paths) is str:
+        img_paths = string_to_list(img_paths)
+    img = load_nifti_file(img_paths[0])
+    img_shape = np.shape(img)
+    for img_path in img_paths:
+        fig = plt.figure()
+        ax = plt.Axes(fig, [0., 0., 1., 1.])
+        ax.set_axis_off()
+        fig.add_axes(ax)
+
+        img = load_nifti_file(img_path)
+    
+        frames = []
+        for index in range(img_shape[-1]):
+            frame = plt.imshow(img[:, :, index], aspect='auto', animated=True)
+            # plt.axis('off')
+            frames.append([frame])
+        
+        anim = animation.ArtistAnimation(fig, frames, interval=interval)
+
+
+        path_to_anim_save = os.path.join(save_path, os.path.split(img_path)[-1].split('.')[0]+'.gif')
+        
+        anim.save(path_to_anim_save)
+        print("Animation saved to:", path_to_anim_save)
+
+    
+def tile_slices(img_paths, save_path="", fname=None, slice_inds=None, col_titles=None):
+    if type(img_paths) is str:
+        img_paths = string_to_list(img_paths)
+    img = load_nifti_file(img_paths[0])
+    img_shape = np.shape(img)
+    
+    if slice_inds is None:
+        slice_inds = [round(np.random.rand() * (img_shape[-1])-1)]
+    
+    if col_titles is None:
+        col_titles = [os.path.split(img_path)[-1].split('.')[0] for img_path in img_paths]
+    
+    num_inds = len(slice_inds)
+    num_imgs = len(img_paths)
+    
+    subplot_mat = np.array(np.arange(num_inds * num_imgs) +1).reshape(num_inds, num_imgs)
+    
+    plt.figure(figsize=(num_imgs*2, num_inds*2))
+    
+    imgs = []
+    for img_path in img_paths:
+        img = load_nifti_file(img_path)
+        imgs.append(img)
+    
+    for img, col_num in zip(imgs, range(num_imgs)):
+        for index, row_num in zip(slice_inds, range(num_inds)):
+            plt.subplot(num_inds, num_imgs, subplot_mat[row_num, col_num])
+            plt.imshow(img[:, :, index])
+            plt.axis('off')
+            if row_num-0<1e-3:
+                plt.title(col_titles[col_num])
+    
+    if fname is None:
+        fname = 'visualisation.png'
+    save_fig_to = os.path.join(save_path, fname)
+    plt.savefig(save_fig_to)
+    print('Plot saved to:', save_fig_to)
+    
+    
+    
+def gif_warp(img_paths, ddf_path, slice_inds=None, num_interval=100, interval=50, save_path=""):
+    if type(img_paths) is str:
+        img_paths = string_to_list(img_paths)
+    
+    image = load_nifti_file(img_paths[0])
+    img_shape = np.shape(image)
+    
+    if slice_inds is None:
+        slice_inds = [round(np.random.rand() * (img_shape[-1])-1)]
+    
+    
[…]
