### 2020-10-04 c2e8d8d
Issue #166: add tiled gif functionality
