### 2020-10-04 01b37f5
Issue #166: add docstrings
