### 2020-10-04 29009ce
Issue #166: make corrections to docstrings
