### 2020-10-04 9a4d83b
Issue #166: fix formatting
