+++ deepreg/vis.py
@@ -4,11 +4,12 @@ command line interfacce
 reuires ffmpeg wirter to write gif files
 """
 
-import numpy as np
-import matplotlib.pyplot as plt
-import matplotlib.animation as animation
-import os
 import argparse
+import os
+
+import matplotlib.animation as animation
+import matplotlib.pyplot as plt
+import numpy as np
 import numpy.matlib
 
 from deepreg.dataset.loader.nifti_loader import load_nifti_file
@@ -16,52 +17,54 @@ from deepreg.model.layer_util import warp_image_ddf
 
 
 def string_to_list(string):
-    '''
+    """
     converts a comma separated string to a list of strings
     also removes leading or trailing spaces from each element in list
-    
+
     :param string: string which is to be converted to list
-    '''
-    return [elem.strip() for elem in string.split(',')]
+    """
+    return [elem.strip() for elem in string.split(",")]
+
 
 def gif_slices(img_paths, save_path="", interval=50):
-    '''
+    """
     Generates and saves gif of slices of image
     supports multiple images to generate multiple gif files
-    
+
     :param img_paths : list or comma separated string of image paths
     :param save_path : path to directory where visualisation/s is/are to be saved
     :param interval : time in miliseconds between frames of gif
-    '''
+    """
     if type(img_paths) is str:
         img_paths = string_to_list(img_paths)
     img = load_nifti_file(img_paths[0])
     img_shape = np.shape(img)
     for img_path in img_paths:
         fig = plt.figure()
-        ax = plt.Axes(fig, [0., 0., 1., 1.])
+        ax = plt.Axes(fig, [0.0, 0.0, 1.0, 1.0])
         ax.set_axis_off()
         fig.add_axes(ax)
 
         img = load_nifti_file(img_path)
-    
+
         frames = []
         for index in range(img_shape[-1]):
-            frame = plt.imshow(img[:, :, index], aspect='auto', animated=True)
+            frame = plt.imshow(img[:, :, index], aspect="auto", animated=True)
             # plt.axis('off')
             frames.append([frame])
-        
+
         anim = animation.ArtistAnimation(fig, frames, interval=interval)
 
+        path_to_anim_save = os.path.join(
+            save_path, os.path.split(img_path)[-1].split(".")[0] + ".gif"
+        )
 
-        path_to_anim_save = os.path.join(save_path, os.path.split(img_path)[-1].split('.')[0]+'.gif')
-        
         anim.save(path_to_anim_save)
         print("Animation saved to:", path_to_anim_save)
 
-    
+
 def tile_slices(img_paths, save_path="", fname=None, slice_inds=None, col_titles=None):
-    '''
+    """
     Generates a tiled plot of muliple images for multiple slices in the image
     rows are different slices and columns are different images
 
@@ -70,49 +73,54 @@ def tile_slices(img_paths, save_path="", fname=None, slice_inds=None, col_titles
     :param fname : file name with extension to save visualisation to
     :param slice_inds : list of slice indices to plot for each image
     :param col_titles : titles for each column, if None then inferred from file names
-    '''
+    """
     if type(img_paths) is str:
         img_paths = string_to_list(img_paths)
     img = load_nifti_file(img_paths[0])
     img_shape = np.shape(img)
-    
+
     if slice_inds is None:
-        slice_inds = [round(np.random.rand() * (img_shape[-1])-1)]
-    
+        slice_inds = [round(np.random.rand() * (img_shape[-1]) - 1)]
+
     if col_titles is None:
-        col_titles = [os.path.split(img_path)[-1].split('.')[0] for img_path in img_paths]
-    
+        col_titles = [
+            os.path.split(img_path)[-1].split(".")[0] for img_path in img_paths
[…]
