### 2020-10-04 DeepRegNet/DeepReg#429 @ethanwharris
[JOSS Review] State of the field
Hi, I'm reviewing the DeepReg paper on JOSS (openjournals/joss-reviews/issues/2705). In order to satisfy the state of the field check in the software paper review criteria, it would be good if you could include some discussion of other libraries that are focussed on medical image registration with deep learning. If there are no other libraries that provide similar functionaility then it would be sufficient to state this in the paper. However, it seems that some brief comparison to [AIRLab](https://github.com/airlab-unibas/airlab) would be appropriate.
