### 2020-10-04 DeepRegNet/DeepReg#430 @ethanwharris
[JOSS Review] Referencing TensorFlow
_Part of the review for the DeepReg paper on JOSS (openjournals/joss-reviews/issues/2705)._

Since DeepReg is built on TensorFlow, it would be valuable to state this in the paper and provide a reference to the [TensorFlow white paper](https://www.tensorflow.org/about/bib).
