### 2020-10-04 DeepRegNet/DeepReg#429 @mathpluscode
> AIRLab

Hi @ethanwharris, thank you for the suggestion! Sure, we will include AIRLab in our paper.
