### 2020-10-04 7187713
Issue #430: cite TensorFlow white paper
