### 2020-10-04 0326476
Issue #429: add references of classical registration packages
