+++ docs/joss_paper/paper.bib
@@ -262,3 +262,31 @@
   journal={arXiv preprint arXiv:1902.09063},
   year={2019}
 }
+
+@article{modat2010fast,
+  title={Fast free-form deformation using graphics processing units},
+  author={Modat, Marc and Ridgway, Gerard R and Taylor, Zeike A and Lehmann, Manja and Barnes, Josephine and Hawkes, David J and Fox, Nick C and Ourselin, S{\'e}bastien},
+  journal={Computer methods and programs in biomedicine},
+  volume={98},
+  number={3},
+  pages={278--284},
+  year={2010},
+  publisher={Elsevier},
+  doi={10.1016/j.cmpb.2009.09.002}
+}
+
+@article{marstal2016simpleelastix,
+  title={SimpleElastix: A user-friendly, multi-lingual library for medical image registration},
+  author={Marstal, Kasper and Berendsen, Floris and Staring, Marius and Klein, Stefan},
+  booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition workshops},
+  pages={134--142},
+  year={2016},
+  doi={10.1109/CVPRW.2016.78}
+}
+
+@article{sandkuhler2018airlab,
+  title={AirLab: autograd image registration laboratory},
+  author={Sandk{\"u}hler, Robin and Jud, Christoph and Andermatt, Simon and Cattin, Philippe C},
+  journal={arXiv preprint arXiv:1806.09907},
+  year={2018}
+}
