+++ docs/joss_paper/paper.md
@@ -74,7 +74,7 @@ Image fusion is a fundamental task in medical image analysis and computer-assist
 A submission to the MICCAI Educational Challenge has utilised the DeepReg code and demos to explore the link between classical algorithms and deep-learning-based methods [@brown2020introduction], while a recently published research work investigated temporal changes in prostate cancer imaging, by using a longitudinal registration adapted from the DeepReg code [@yang2020longitudinal].
 
 # Statement of need
-Currently, popular packages focusing on deep learning methods for medical imaging, such as NiftyNet [@gibson2018niftynet] and MONAI (https://monai.io/), do not support image registration. Other open-source projects implementing specific published algorithms, such as the VoxelMorph [@balakrishnan2019voxelmorph], are seldom tested or designed for general research and education purposes. Therefore an open-sourced project focusing on image registration with deep learning is much needed.
+Currently, popular packages focusing on deep learning methods for medical imaging, such as NiftyNet [@gibson2018niftynet] and MONAI (https://monai.io/), do not support image registration. The existing open-sourced registration projects either implement only specific published algorithms, such as the VoxelMorph [@balakrishnan2019voxelmorph], or focus only on classical methods, such as NiftiReg [@modat2010fast], SimpleElastix [@marstal2016simpleelastix] and AirLab [@sandkuhler2018airlab]. Therefore an open-sourced project focusing on image registration with deep learning is much needed for general research and education purposes.
 
 # Implementation
 DeepReg implements a framework for unsupervised learning [@de2019deep; @balakrishnan2019voxelmorph], weakly-supervised learning [@hu2018label; @hu2018weakly] and their combinations and variants, e.g. [@hu2019conditional]. Many options are included for major components of these approaches, such as different image- and label dissimilarity functions, transformation models [@ashburner2007fast; @vercauteren2009diffeomorphic; @hill2001medical], deformation regularisation [@rueckert1999nonrigid] and different neural network architectures [@hu2018weakly; @he2016deep; @simonyan2014very]. The provided dataset loaders adopt staged random sampling strategy to ensure unbiased learning from groups, images and labels [@hu2018weakly; @yang2020longitudinal]. These algorithmic components together with the flexible dataset loaders are building blocks of many other registration tasks, such as group-wise registration and morphological template construction [@dalca2019learning; @siebert2020deep; @luo2020mvmm].
