### 2020-10-04 DeepRegNet/DeepReg#430 @mathpluscode
@ethanwharris We've updated the paper regarding this issue. Once #434 is resolved, we will update the pdf in the JOSS issue. Thanks again for kind suggestions.
