### 2020-10-04 9154da9
Issue #166: add unit tests for vis
