### 2020-10-04 3a64655
Issue #166: add doc for vis tool
