+++ docs/source/docs/visualisation_tool.md
@@ -1,0 +1,162 @@
+# Visualisation tool
+
+DeepReg provides a visuaisation tool which allows the user to generate various
+visualisations from nifti images. The tool is compatible with outputs from
+deepreg_predict as well as with other nifti images.
+
+The visualisation tool currently offers four unique functionalities which are described
+below. Each of these can be accessed using the `deepreg_vis` CLI tool. Examples are
+provided for each functionality but these can only be run after changing directory to
+the root folder which is `DeepReg`, this is to ensure that the test images are
+accessible. The functionalities can also be accessed by importing functions into a
+python script and this method is described in the last section of this document.
+
+## General arguments for CLI tool
+
+Some general arguments for the `deepreg_vis` CLI tool are as follows:
+
+- `-m` or `--mode`: This specifies which mode to use to generate the visualisation. See
+  below for available modes.
+- `-i` or `--image-paths`: This is the path of the image or images that need to be used
+  to generate the visualisation. Multiple paths can be passed using a comma separated
+  string.
+- `-s` or `--save-path`: This is the path to the directory where the visualisation will
+  be saved. The name of the visualisation is auto generated, however, if `--fname`
+  argument is available for a specific mode, then a custom filename can be specified.
+  (default results in saving to current directory)
+
+## GIF over image slices
+
+This creates an animation which is an iteration through the image slices of a 3D image.
+
+In addition to the general arguments, the additional arguments applicable to this mode
+are:
+
+- `--interval`: This argument is optional and can be used to specify the time, in
+  milliseconds, between successive frames of an animation. (default=50)
+
+It can be accessed by passing `--mode 0` or `-m 0` to the CLI tool.
+
+The output will be a file with the same name as the original with a `.gif` extension. If
+multiple image paths are passed in the `-i` or `--image-paths` argument then multiple
+`.gif` files will be generated, one for each unique image path passed.
+
+A simple example, which takes an image path and saves a `.gif` animation, is shown
+below:
+
+```
+deepreg_vis --m 0 -i ./data/test/nifti/unit_test/moving_image.nii.gz -s logs
+```
+
+## GIF that shows warping
+
+This functionality produces an animation showing warping for a single image slice using
+a ddf.
+
+In addition to the general arguments, the additional arguments applicable to this mode
+are:
+
+- `--ddf-path`: This argument is required for this mode and specifies the path of the
+  ddf to use for warping the image.
+- `--slice-inds`: This argument is optional and can be used to specify the indexes to be
+  used to generate the visualisation. Multiple can be passed by using a comma separated
+  string. (default results in a random slice being used)
+- `--interval`: This argument is optional and can be used to specify the time, in
+  milliseconds, between successive frames of an animation. (default=50)
+- `--num-interval`: The number of intervals to use for warping. (default=100)
+
+It can be accessed by passing `--mode 1` or `-m 1` to the CLI tool.
+
+The output will be a file with the slice number appended to the original file name, with
+a `.gif` extension. For example if a file is named `moving_image.nii.gz` and slice
+number 2 and 3 three are chosen, the files produced will be `moving_image_slice_2.gif`
+and `moving_image_slice_3.gif`.If multiple image paths are passed in the `-i` or
+`--image-paths` argument then multiple `.gif` files will be generated, if multiple slice
+indexes are specified then multiple `.gif` files are generated for each unique image
+path passed.
+
+A simple example, which takes an image, slice indexes and a ddf path and saves a `.gif`
+animation for each slice, is shown below:
+
+```
[…]
