### 2020-10-04 5a404c5
Issue #166: modify visualisation tool doc
