### 2020-10-04 f19e0a8
Issue #436: do not normalize DDF/DVF before saving them
