+++ deepreg/predict.py
@@ -16,7 +16,6 @@ import tensorflow as tf
 import deepreg.model.layer_util as layer_util
 import deepreg.model.optimizer as opt
 import deepreg.parser as config_parser
-from deepreg.dataset.loader.util import normalize_array
 from deepreg.model.network.build import build_model
 from deepreg.util import (
     build_dataset,
@@ -167,10 +166,9 @@ def predict_on_dataset(
             names = ["ddf", "dvf"]
             for arr, name in zip(arrs, names):
                 if arr is not None:
-                    arr = normalize_array(arr=arr[sample_index, :, :, :])
                     save_array(
                         save_dir=label_dir if conditional else pair_dir,
-                        arr=arr,
+                        arr=arr[sample_index, :, :, :],
                         name=name,
                         gray=False,
                         save_nifti=save_nifti,
