### 2020-10-05 64e3f43
Issue #436: save_array normalizes arrays before saving png
