+++ test/unit/test_util.py
@@ -104,7 +104,7 @@ def test_save_array():
     name = "3d_tf"
     out_dir = os.path.join(save_dir, name)
     arr = tf.random.uniform(shape=(2, 3, 4))
-    save_array(save_dir=save_dir, arr=arr, name=name, gray=True)
+    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True)
     assert get_num_pngs_in_dir(out_dir) == 4
     assert get_num_niftis_in_dir(save_dir) == 1
     shutil.rmtree(out_dir)
@@ -114,7 +114,7 @@ def test_save_array():
     name = "4d_tf"
     out_dir = os.path.join(save_dir, name)
     arr = tf.random.uniform(shape=(2, 3, 4, 3))
-    save_array(save_dir=save_dir, arr=arr, name=name, gray=True)
+    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True)
     assert get_num_pngs_in_dir(out_dir) == 4
     assert get_num_niftis_in_dir(save_dir) == 1
     shutil.rmtree(out_dir)
@@ -124,7 +124,7 @@ def test_save_array():
     name = "3d_np"
     out_dir = os.path.join(save_dir, name)
     arr = np.random.rand(2, 3, 4)
-    save_array(save_dir=save_dir, arr=arr, name=name, gray=True)
+    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True)
     assert get_num_pngs_in_dir(out_dir) == 4
     assert get_num_niftis_in_dir(save_dir) == 1
     shutil.rmtree(out_dir)
@@ -134,7 +134,7 @@ def test_save_array():
     name = "4d_np"
     out_dir = os.path.join(save_dir, name)
     arr = np.random.rand(2, 3, 4, 3)
-    save_array(save_dir=save_dir, arr=arr, name=name, gray=True)
+    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True)
     assert get_num_pngs_in_dir(out_dir) == 4
     assert get_num_niftis_in_dir(save_dir) == 1
     shutil.rmtree(out_dir)
@@ -144,7 +144,7 @@ def test_save_array():
     name = "4d_np"
     out_dir = os.path.join(save_dir, name)
     arr = np.random.rand(2, 3, 4, 3)
-    save_array(save_dir=save_dir, arr=arr, name=name, gray=True, save_nifti=False)
+    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True, save_nifti=False)
     assert get_num_pngs_in_dir(out_dir) == 4
     assert get_num_niftis_in_dir(save_dir) == 0
     shutil.rmtree(out_dir)
@@ -154,7 +154,7 @@ def test_save_array():
     out_dir = os.path.join(save_dir, name)
     arr = np.random.rand(2, 3, 4, 3)
     assert not os.path.exists(out_dir)
-    save_array(save_dir=save_dir, arr=arr, name=name, gray=True, save_png=False)
+    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True, save_png=False)
     assert not os.path.exists(out_dir)
     assert get_num_niftis_in_dir(save_dir) == 1
     os.remove(os.path.join(save_dir, name + ".nii.gz"))
@@ -170,11 +170,11 @@ def test_save_array():
     os.makedirs(save_dir, exist_ok=True)
     nib.save(img=nib.Nifti2Image(arr1, affine=np.eye(4)), filename=nifti_file_path)
     # save arr2 without overwrite
-    save_array(save_dir=save_dir, arr=arr1, name=name, gray=True, overwrite=False)
+    save_array(save_dir=save_dir, arr=arr1, name=name, normalize=True, overwrite=False)
[…]
