### 2020-10-05 b933fc8
Issue #436: refactor save_array tests
