+++ test/unit/test_util.py
@@ -85,114 +85,105 @@ def test_build_log_dir(log_root: str, log_dir: str):
         assert tail == log_dir
 
 
-def test_save_array():
-    """
-    Test save_array by testing different shapes and count output files
-    """
-
-    def get_num_pngs_in_dir(dir_paths):
-        return len([x for x in os.listdir(dir_paths) if x.endswith(".png")])
-
-    def get_num_niftis_in_dir(dir_paths):
-        return len([x for x in os.listdir(dir_paths) if x.endswith(".nii.gz")])
-
+class TestSaveArray:
     save_dir = "logs/test_util_save_array"
-    if os.path.exists(save_dir):
-        shutil.rmtree(save_dir)
+    arr_name = "arr"
+    png_dir = os.path.join(save_dir, arr_name)
+    dim_err_msg = "arr must be 3d or 4d numpy array or tf tensor"
+    ch_err_msg = "4d arr must have 3 channels as last dimension"
 
-    # test 3D tf tensor
-    name = "3d_tf"
-    out_dir = os.path.join(save_dir, name)
-    arr = tf.random.uniform(shape=(2, 3, 4))
-    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True)
-    assert get_num_pngs_in_dir(out_dir) == 4
-    assert get_num_niftis_in_dir(save_dir) == 1
-    shutil.rmtree(out_dir)
-    os.remove(os.path.join(save_dir, name + ".nii.gz"))
+    def setup_method(self, method):
+        if os.path.exists(self.save_dir):
+            shutil.rmtree(self.save_dir)
 
-    # test 4D tf tensor
-    name = "4d_tf"
-    out_dir = os.path.join(save_dir, name)
-    arr = tf.random.uniform(shape=(2, 3, 4, 3))
-    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True)
-    assert get_num_pngs_in_dir(out_dir) == 4
-    assert get_num_niftis_in_dir(save_dir) == 1
-    shutil.rmtree(out_dir)
-    os.remove(os.path.join(save_dir, name + ".nii.gz"))
+    def teardown_method(self, method):
+        if os.path.exists(self.save_dir):
+            shutil.rmtree(self.save_dir)
 
-    # test 3D np tensor
-    name = "3d_np"
-    out_dir = os.path.join(save_dir, name)
-    arr = np.random.rand(2, 3, 4)
-    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True)
-    assert get_num_pngs_in_dir(out_dir) == 4
-    assert get_num_niftis_in_dir(save_dir) == 1
-    shutil.rmtree(out_dir)
-    os.remove(os.path.join(save_dir, name + ".nii.gz"))
+    @staticmethod
+    def get_num_files_in_dir(dir_path: str, suffix: str):
+        if os.path.exists(dir_path):
+            return len([x for x in os.listdir(dir_path) if x.endswith(suffix)])
+        return 0
 
-    # test 4D np tensor
-    name = "4d_np"
-    out_dir = os.path.join(save_dir, name)
-    arr = np.random.rand(2, 3, 4, 3)
-    save_array(save_dir=save_dir, arr=arr, name=name, normalize=True)
-    assert get_num_pngs_in_dir(out_dir) == 4
-    assert get_num_niftis_in_dir(save_dir) == 1
-    shutil.rmtree(out_dir)
-    os.remove(os.path.join(save_dir, name + ".nii.gz"))
+    @pytest.mark.parametrize(
+        "arr",
+        [
+            tf.random.uniform(shape=(2, 3, 4)),
+            tf.random.uniform(shape=(2, 3, 4, 3)),
+            np.random.rand(2, 3, 4),
[…]
