### 2020-10-06 a38accc
Issue #436: refactor tests related to warp
