+++ test/unit/test_warp.py
@@ -1,5 +1,7 @@
 import os
 
+import pytest
+
 from deepreg.warp import main
 
 
@@ -8,42 +10,25 @@ def test_warp():
     pass
 
 
-def test_main():
-    """
-    Test main by checking the output file existance
-    """
+class TestMain:
     image_path = "./data/test/nifti/unit_test/moving_image.nii.gz"
     ddf_path = "./data/test/nifti/unit_test/ddf.nii.gz"
 
-    # custom output path with correct suffix
-    out_path = "logs/test_warp/out.nii.gz"
-    main(args=["--image", image_path, "--ddf", ddf_path, "--out", out_path])
-    assert os.path.isfile(out_path)
-    os.remove(out_path)
-
-    # custom output path without correct suffix
-    out_path = "logs/test_warp/out.h5"
-    main(args=["--image", image_path, "--ddf", ddf_path, "--out", out_path])
-    out_path = "logs/test_warp/warped.nii.gz"
-    assert os.path.isfile(out_path)
-    os.remove(out_path)
-
-    # custom output path without correct suffix
-    out_path = "logs/test_warp/"
-    main(args=["--image", image_path, "--ddf", ddf_path, "--out", out_path])
-    out_path = "logs/test_warp/warped.nii.gz"
-    assert os.path.isfile(out_path)
-    os.remove(out_path)
-
-    # custom output path without correct suffix
-    out_path = "logs/test_warp"
-    main(args=["--image", image_path, "--ddf", ddf_path, "--out", out_path])
-    out_path = "logs/warped.nii.gz"
-    assert os.path.isfile(out_path)
-    os.remove(out_path)
-
-    # custom output path
-    out_path = "warped.nii.gz"
-    main(args=["--image", image_path, "--ddf", ddf_path, "--out", ""])
-    assert os.path.isfile(out_path)
-    os.remove(out_path)
+    @pytest.mark.parametrize(
+        "out_path,expected_path",
+        [
+            ["logs/test_warp/out.nii.gz", "logs/test_warp/out.nii.gz"],
+            ["logs/test_warp/out.h5", "logs/test_warp/warped.nii.gz"],
+            ["logs/test_warp/", "logs/test_warp/warped.nii.gz"],
+            ["", "warped.nii.gz"],
+        ],
+    )
+    def test_main(self, out_path: str, expected_path: str):
+        """
+        Integration test by checking the output file exists
+        """
+        main(
+            args=["--image", self.image_path, "--ddf", self.ddf_path, "--out", out_path]
+        )
+        assert os.path.isfile(expected_path)
+        os.remove(expected_path)
