### 2020-10-06 0796c8d
Issue #420: add environment.yml
