+++ requirements.txt
@@ -1,25 +1,24 @@
-h5py
-numpy>=1.16
-pandas
-nibabel
-pyyaml>=5.3.1
-matplotlib
 argparse
-tqdm
-tensorflow==2.2
-pytest>=4.6
-pytest-cov
-pytest-dependency
-pre-commit
-seed-isort-config
-isort
 black>=20.8b1
 flake8
-testfixtures
-notebook
-simple_http_server
-sphinx==3.2.1
-sphinx_rtd_theme
-sphinx-tabs
+h5py
+isort
 m2r2
+matplotlib
+nibabel
+notebook
+numpy
+pandas
+pre-commit
+pytest
+pytest-cov
+pytest-dependency
+pyyaml>=5.3.1
+seed-isort-config
+sphinx==3.2.1
+sphinx-rtd-theme
+sphinx-tabs
 sphinx-notfound-page
+tensorflow-gpu==2.3.0
+testfixtures
+tqdm
