### 2020-10-06 2dbd9b8
Issue #420: add cuda cudnn requirements in environment.yml
