+++ environment.yml
@@ -1,7 +1,10 @@
 name: deepreg
 channels:
   - defaults
+  - conda-forge
 dependencies:
+  - cudatoolkit=10.1
+  - cudnn=7.6
   - python=3.7
   - pip
   - pip:
