### 2020-10-06 0d227bb
Issue #442: correct is_equal definition
