### 2020-10-06 50048ba
Issue #442: fix and refactor normalize_array and its tests
