+++ test/unit/test_dataset_loader_util.py
@@ -7,47 +7,45 @@ Tests for deepreg/dataset/loader/util.py in pytest style
 from test.unit.util import is_equal_np
 
 import numpy as np
+import pytest
 
 import deepreg.dataset.loader.util as util
 
 
-def test_normalize_array():
-    """
-    Test normalize array by checking the output values
-    """
+class TestNormalizeArray:
+    @pytest.mark.parametrize(
+        "arr,expected",
+        [
+            [np.array([0, 1, 2]), np.array([0, 0.5, 1])],
+            [np.array([-2, 0, 1, 2]), np.array([0, 0.5, 0.75, 1])],
+            [np.array([1, 1, 1, 1]), np.array([0, 0, 0, 0])],
+        ],
+    )
+    def test_no_min_no_max(self, arr, expected):
+        got = util.normalize_array(arr=arr)
+        assert is_equal_np(got, expected)
 
-    # no v_min, v_max
-    arr = np.array([0, 1, 2])
-    got = util.normalize_array(arr=arr)
-    expected = np.array([0, 0.5, 1])
-    assert is_equal_np(got, expected)
+    @pytest.mark.parametrize(
+        "arr,v_max,expected",
+        [
+            [np.array([0, 1, 2]), 1, np.array([0, 1, 1])],
+            [np.array([-2, 0, 1, 2]), 3, np.array([0, 0.4, 0.6, 0.8])],
+        ],
+    )
+    def test_no_min(self, arr, v_max, expected):
+        got = util.normalize_array(arr=arr, v_max=v_max)
+        assert is_equal_np(got, expected)
 
-    arr = np.array([-2, 0, 1, 2])
-    got = util.normalize_array(arr=arr)
-    expected = np.array([0, 0.5, 0.75, 1])
-    assert is_equal_np(got, expected)
-
-    # no v_min
-    arr = np.array([0, 1, 2])
-    got = util.normalize_array(arr=arr, v_max=1)
-    expected = np.array([0, 1, 1])
-    assert is_equal_np(got, expected)
-
-    arr = np.array([-2, 0, 1, 2])
-    got = util.normalize_array(arr=arr, v_max=3)
-    expected = np.array([0, 0.4, 0.6, 0.8])
-    assert is_equal_np(got, expected)
-
-    # no v_max
-    arr = np.array([0, 1, 2])
-    got = util.normalize_array(arr=arr, v_min=1)
-    expected = np.array([0, 0, 1])
-    assert is_equal_np(got, expected)
-
-    arr = np.array([-2, 0, 1, 2])
-    got = util.normalize_array(arr=arr, v_min=-3)
-    expected = np.array([0.2, 0.6, 0.8, 1])
-    assert is_equal_np(got, expected)
+    @pytest.mark.parametrize(
+        "arr,v_min,expected",
+        [
+            [np.array([0, 1, 2]), 1, np.array([0, 0, 1])],
+            [np.array([-2, 0, 1, 2]), -3, np.array([0.2, 0.6, 0.8, 1])],
+        ],
+    )
+    def test_no_max(self, arr, v_min, expected):
+        got = util.normalize_array(arr=arr, v_min=v_min)
+        assert is_equal_np(got, expected)
 
 
 def test_remove_prefix_suffix():
