### 2020-10-06 8836b34
Issue #442: use tf.math.divide_no_nan instead of +eps and fix tests
