### 2020-10-07 8e45fc4
Issue #442: use tf.math.divide_no_nan for GMI loss and fix tests
