+++ test/unit/util.py
@@ -12,9 +12,11 @@ def is_equal_np(x: (np.ndarray, List), y: (np.ndarray, List)) -> bool:
 
 
 def is_equal_tf(
-    x: (tf.Tensor, np.ndarray, List), y: (tf.Tensor, np.ndarray, List)
+    x: (tf.Tensor, np.ndarray, List),
+    y: (tf.Tensor, np.ndarray, List),
+    atol: float = 1.0e-8,
 ) -> bool:
     """return true if two tf tensors are nearly equal"""
     x = tf.cast(x, dtype=tf.float32).numpy()
     y = tf.cast(y, dtype=tf.float32).numpy()
-    return np.all(np.isclose(x, y))
+    return np.all(np.isclose(x, y, atol=atol))
