+++ test/unit/test_loss_image.py
@@ -13,70 +13,56 @@ import tensorflow as tf
 import deepreg.model.loss.image as image
 
 
-def test_dissimilarity_fn():
+class TestDissimilarityFn:
     """
-    Testing computed dissimilarity function by comparing to precomputed, the dissimilarity function can be either normalized cross correlation or sum square error function.
+    Testing computed dissimilarity function by comparing to precomputed,
+    the dissimilarity function can be either normalized cross correlation or sum square error function.
     """
 
-    # lncc diff images
-    tensor_true = np.array(range(12)).reshape((2, 1, 2, 3))
-    tensor_pred = 0.6 * np.ones((2, 1, 2, 3))
-    tensor_true = tf.convert_to_tensor(tensor_true, dtype=tf.float32)
-    tensor_pred = tf.convert_to_tensor(tensor_pred, dtype=tf.float32)
+    y_true = tf.constant(np.array(range(12)).reshape((2, 1, 2, 3)), dtype=tf.float32)
+    y_pred = 0.6 * tf.ones((2, 1, 2, 3), dtype=tf.float32)
 
-    name_ncc = "lncc"
-    get_ncc = image.dissimilarity_fn(tensor_true, tensor_pred, name_ncc)
-    expect_ncc = [-0.68002254, -0.9608879]
+    def test_lncc(self):
+        got = image.dissimilarity_fn(self.y_true, self.y_pred, "lncc")
+        expected = [-0.68002254, -0.9608879]
+        assert is_equal_tf(got, expected)
 
-    assert is_equal_tf(get_ncc, expect_ncc)
+        got = image.dissimilarity_fn(self.y_pred, self.y_pred, "lncc")
+        expected = [-1, -1]
+        assert is_equal_tf(got, expected)
 
-    # ssd diff images
-    tensor_true1 = np.zeros((2, 1, 2, 3))
-    tensor_pred1 = 0.6 * np.ones((2, 1, 2, 3))
-    tensor_true1 = tf.convert_to_tensor(tensor_true1, dtype=tf.float32)
-    tensor_pred1 = tf.convert_to_tensor(tensor_pred1, dtype=tf.float32)
+    def test_ssd(self):
+        got = image.dissimilarity_fn(self.y_pred * 0, self.y_pred, "ssd")
+        expected = [0.36, 0.36]
+        assert is_equal_tf(got, expected)
 
-    name_ssd = "ssd"
-    get_ssd = image.dissimilarity_fn(tensor_true1, tensor_pred1, name_ssd)
-    expect_ssd = [0.36, 0.36]
+        got = image.dissimilarity_fn(self.y_pred, self.y_pred, "ssd")
+        expected = [0, 0]
+        assert is_equal_tf(got, expected)
 
-    assert is_equal_tf(get_ssd, expect_ssd)
+    def test_gmi(self):
+        # TODO gmi diff images
+        # gmi same image
+        t = tf.ones([4, 3, 3, 3])
+        get_zero_similarity_gmi = image.dissimilarity_fn(t, t, "gmi")
+        assert is_equal_tf(get_zero_similarity_gmi, [0, 0, 0, 0], atol=1.0e-7)
 
-    # TODO gmi diff images
-
-    # lncc same image
-    get_zero_similarity_ncc = image.dissimilarity_fn(
-        tensor_pred1, tensor_pred1, name_ncc
-    )
-    assert is_equal_tf(get_zero_similarity_ncc, [-1, -1])
-
-    # ssd same image
-    get_zero_similarity_ssd = image.dissimilarity_fn(
-        tensor_true1, tensor_true1, name_ssd
-    )
-    assert is_equal_tf(get_zero_similarity_ssd, [0, 0])
-
-    # gmi same image
[…]
