### 2020-10-07 cb8d40e
Issue #442: increase atol for GMI
Tests on PT failed.
