### 2020-10-07 a46c975
Issue #442: fix failed test with GPU for tf.nn.conv3d
filters and input shapes didn't match
