### 2020-10-07 7abdcb5
Issue #166: correct typo in visualisation tool doc
