+++ docs/source/docs/visualisation_tool.md
@@ -51,7 +51,7 @@ A simple example, which takes an image path and saves a `.gif` animation, is sho
 below:
 
 ```
-deepreg_vis --m 0 -i ./data/test/nifti/unit_test/moving_image.nii.gz -s logs
+deepreg_vis -m 0 -i ./data/test/nifti/unit_test/moving_image.nii.gz -s logs
 ```
 
 ## GIF that shows warping
@@ -85,7 +85,7 @@ A simple example, which takes an image, slice indexes and a ddf path and saves a
 animation for each slice, is shown below:
 
 ```
-deepreg_vis --m 1 -i ./data/test/nifti/unit_test/moving_image.nii.gz --ddf-path "./data/test/nifti/unit_test/ddf.nii.gz" --slice-inds '2,3' -s logs
+deepreg_vis -m 1 -i ./data/test/nifti/unit_test/moving_image.nii.gz --ddf-path "./data/test/nifti/unit_test/ddf.nii.gz" --slice-inds '2,3' -s logs
 ```
 
 ## Plot of image slices
@@ -114,7 +114,7 @@ A simple example, which takes three images and three slice indexes and saves a `
 file, is shown below (this will create a plot with 3 columns and 3 rows):
 
 ```
-deepreg_vis --m 1 -i './data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz' --slice-inds '2,3,4' -s logs
+deepreg_vis -m 1 -i './data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz' --slice-inds '2,3,4' -s logs
 ```
 
 ## Tiled GIF over image slices
@@ -140,7 +140,7 @@ A simple example, which takes four images and creates an animation over the slic
 the four images in a tiled manner, is shown below:
 
 ```
-deepreg_vis --m 1 -i './data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz' --size '2,2' -s logs
+deepreg_vis -m 1 -i './data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz, ./data/test/nifti/unit_test/moving_image.nii.gz' --size '2,2' -s logs
 ```
 
 ## Using the same functionality by importing python functions
