### 2020-10-07 b3496b9
Issue #166: correct minor typo in vis tool doc
