### 2020-10-07 0263830
Issue #166: correct typo in docstrings of vis.py
