### 2020-10-08 cef9e79
Issue #166 #445 minor text updates with adding pointer in cli section
