### 2020-10-08 493a2f5
Issue #166: add unit test for vis.py to cover lines missing in codecov
