### 2020-10-09 fa0ff5a
Issue #442: batch_size in config means for per GPU
