### 2020-10-09 4a6fe92
Issue #442: use tf func directly for getting gpu devices
