+++ deepreg/train.py
@@ -12,7 +12,7 @@ import tensorflow as tf
 import deepreg.model.optimizer as opt
 import deepreg.parser as config_parser
 from deepreg.model.network.build import build_model
-from deepreg.util import build_dataset, build_log_dir, get_available_gpus
+from deepreg.util import build_dataset, build_log_dir
 
 
 def build_config(
@@ -45,7 +45,8 @@ def build_config(
     config_parser.save(config=config, out_dir=log_dir)
 
     # batch_size in original config corresponds to batch_size per GPU
-    config["train"]["preprocess"]["batch_size"] *= max(len(get_available_gpus()), 1)
+    gpus = tf.config.experimental.list_physical_devices("GPU")
+    config["train"]["preprocess"]["batch_size"] *= max(len(gpus), 1)
 
     return config, log_dir
 
