### 2020-10-09 643fc70
Issue #436: update docs related to png saving.
