### 2020-10-10 21811bb
Issue #442: cover all code in image.py
