### 2020-10-10 44beb8e
Issue #442: cover all code in label.py
