### 2020-10-10 90b75ac
Issue #420: add tensorboard_plugin_profile into requirements
