### 2020-10-10 41d785b
Issue #420: update tf to 2.3.1
