### 2020-10-10 28889e0
Issue #420: fix np version to 1.18.5 as tf2.3 is not supporting 1.19
