+++ requirements.txt
@@ -7,7 +7,7 @@ m2r2
 matplotlib
 nibabel
 notebook
-numpy
+numpy==1.18.5
 pandas
 pre-commit
 pytest
