### 2020-10-10 105ded6
Issue #420: add cpu version of env.yml
