+++ environment_cpu.yml
@@ -2,10 +2,7 @@ name: deepreg
 channels:
   - defaults
 dependencies:
-  - cudatoolkit=10.1
-  - cudnn=7.6
   - python=3.7
   - pip
   - pip:
-      - tensorflow-gpu==2.3.1
       - -r file:requirements.txt
