### 2020-10-10 a9763fb
Issue #420: add docker file and update docs
