### 2020-10-10 c0b8579
Issue #420: remove the redundant file
