+++ docs/source/tutorial/run_cluster.md
@@ -1,10 +1,11 @@
-# Running DeepReg on the Cluster
+# Running DeepReg Remotely (on the Cluster)
+
 This tutorial gives an example of how to run DeepReg remotely (e.g. on a cluster).  Our example is specific to the UCL cluster, which has operating system CentOS 7 (similar to Ubuntu), with job scheduler Sun Grid Engine (SGE).  More information on the specific configuration at UCL is available [here](https://hpc.cs.ucl.ac.uk/job-submission/).
 
-# Installing the Environment
+## Installing the Environment
 Install the environment in the cluster, as described [here](../../../README.md).  In the case you do not have root access (as the case of the UCL cluster), you might need to use pip -u option to install the requirements.txt file.
 
-# Example Script
+## Example Script
 Below is the submission script for running quick start example [here](../getting_started/quick_start.md).  Change `<DeepReg_dir>` in the script to the remote DeepReg repo location and save the below code in a `<your_name>.qsub`.  Submit the job with `qsub <your_name>.qsub` and check the status of the job with `qstat`, the saved stdout and stderr is in `<DeepReg_dir>/logs/`.
 
 ```
@@ -31,5 +32,5 @@ deepreg_train \
 --log_dir test
 ```
 
-# Contact and Version
-Please contact stefano.blumberg.17@ucl.ac.uk, for information about this questions about the cluster.  Written on 2020-10-11.
+## Contact and Version
+Please contact stefano.blumberg.17@ucl.ac.uk, for information about this questions about the cluster.
