### 2020-10-11 68c4dfb
Issue #413: resolved pre-commit
