### 2020-10-15 be7adfd
Issue #215: Add kernel_type kwarg + implementation of linear and gaussian kernel to LNCC loss
