### 2020-10-15 abcdfec
Issue #215: update docstring
