### 2020-10-15 1b705b0
Issue #215: small correction for tf2.3.0 convolutional syntaxis
