+++ deepreg/model/loss/image.py
@@ -74,25 +74,39 @@ def local_normalized_cross_correlation(
 
     elif kernel_type == "triangular":
         fsize = int((kernel_size + 1) / 2)
+        pad_filter = tf.constant(
+            [
+                [0, 0],
+                [int((fsize - 1) / 2), int((fsize - 1) / 2)],
+                [int((fsize - 1) / 2), int((fsize - 1) / 2)],
+                [int((fsize - 1) / 2), int((fsize - 1) / 2)],
+                [0, 0],
+            ]
+        )
+
         f1 = tf.ones(shape=(1, fsize, fsize, fsize, 1)) / fsize
+        f1 = tf.pad(f1, pad_filter, "CONSTANT")
         f2 = tf.ones(shape=(fsize, fsize, fsize, 1, ch)) / fsize
 
-        pad_filter = [
-            [0, 0],
-            [fsize - 1, fsize - 1],
-            [fsize - 1, fsize - 1],
-            [fsize - 1, fsize - 1],
-            [0, 0],
-        ]
-        filters = tf.nn.conv3d(f1, f2, strides=1, padding=pad_filter)
+        filters = tf.nn.conv3d(f1, f2, strides=[1, 1, 1, 1, 1], padding="SAME")
+        filters = tf.transpose(filters, perm=[1, 2, 3, 4, 0])
         kernel_vol = tf.reduce_sum(filters ** 2)
 
     elif kernel_type == "gaussian":
         mean = (kernel_size - 1) / 2.0
         sigma = kernel_size / 3
 
-        grid_dim = tf.range(0, kernel_size - 1)
-        grid = tf.cast(tf.stack(tf.meshgrid(grid_dim, grid_dim), 0), dtype="float32")
+        grid_dim = tf.range(0, kernel_size)
+        grid = tf.expand_dims(
+            tf.expand_dims(
+                tf.cast(
+                    tf.stack(tf.meshgrid(grid_dim, grid_dim, grid_dim), 0),
+                    dtype="float32",
+                ),
+                axis=-1,
+            ),
+            axis=-1,
+        )
         filters = tf.exp(
             -tf.reduce_sum(tf.square(grid - mean), axis=0) / (2 * sigma ** 2)
         )
@@ -178,7 +192,6 @@ def global_mutual_information(
         tf.reduce_mean(bin_centers[:, :, 1:] - bin_centers[:, :, :-1]) * sigma_ratio
     )  # scalar, sigma in the Gaussian function (weighting function W)
     preterm = 1 / (2 * tf.math.square(sigma))  # scalar
-
     batch, w, h, z, c = y_true.shape
     y_true = tf.reshape(y_true, [batch, w * h * z, 1])  # (batch, nb_voxels, 1)
     y_pred = tf.reshape(y_pred, [batch, w * h * z, 1])  # (batch, nb_voxels, 1)
