### 2020-10-15 86a5c87
Issue #215: small update unit tests for image_loss
