### 2020-10-15 27a58bd
Issue #452: use eps instead of div_no_nan for division
