+++ deepreg/model/loss/image.py
@@ -3,7 +3,7 @@ Module provides different loss functions for calculating the dissimilarities bet
 """
 import tensorflow as tf
 
-EPS = 1.0e-6  # epsilon to prevent NaN
+EPS = tf.keras.backend.epsilon()
 
 
 def dissimilarity_fn(
@@ -97,7 +97,7 @@ def local_normalized_cross_correlation(
     cross = tp_sum - p_avg * t_sum
     t_var = t2_sum - t_avg * t_sum  # std[t] ** 2
     p_var = p2_sum - p_avg * p_sum  # std[p] ** 2
-    ncc = tf.math.divide_no_nan(cross * cross, t_var * p_var)
+    ncc = (cross * cross + EPS) / (t_var * p_var + EPS)
     return tf.reduce_mean(ncc, axis=[1, 2, 3, 4])
 
 
@@ -161,6 +161,5 @@ def global_mutual_information(
     pab /= nb_voxels
 
     # MI: sum(P_ab * log(P_ab/P_aP_b))
-    div = tf.math.divide_no_nan(pab, papb)
-    div = tf.math.maximum(div, EPS)
+    div = (pab + EPS) / (papb + EPS) + EPS
     return tf.reduce_sum(pab * tf.math.log(div), axis=[1, 2])
