### 2020-10-15 029b218
Issue #453: add instruction to get commit hash
