### 2020-10-16 9e2dad6
Issue #452: update EPS in logarithms and denominators (not in numerators)
