+++ deepreg/model/loss/label.py
@@ -138,9 +138,11 @@ def weighted_binary_cross_entropy(
     :param pos_weight: weight of positive class, scalar. Default value is 1
     :return: shape = (batch,)
     """
-    y_pred = tf.clip_by_value(y_pred, EPS, 1 - EPS)
-    loss_pos = tf.reduce_mean(y_true * tf.math.log(y_pred), axis=[1, 2, 3])
-    loss_neg = tf.reduce_mean((1 - y_true) * tf.math.log(1 - y_pred), axis=[1, 2, 3])
+    y_pred = tf.clip_by_value(y_pred, 0, 1)
+    loss_pos = tf.reduce_mean(y_true * tf.math.log(y_pred + EPS), axis=[1, 2, 3])
+    loss_neg = tf.reduce_mean(
+        (1 - y_true) * tf.math.log(1 - y_pred + EPS), axis=[1, 2, 3]
+    )
     return -pos_weight * loss_pos - loss_neg
 
 
@@ -166,7 +168,7 @@ def dice_score(y_true: tf.Tensor, y_pred: tf.Tensor, binary: bool = False) -> tf
     denominator = tf.reduce_sum(y_true, axis=[1, 2, 3]) + tf.reduce_sum(
         y_pred, axis=[1, 2, 3]
     )
-    return (numerator + EPS) / (denominator + EPS)
+    return numerator / (denominator + EPS)
 
 
 def dice_score_generalized(
@@ -201,7 +203,7 @@ def dice_score_generalized(
         (pos_weight + neg_weight) * y_prod - neg_weight * y_sum + neg_weight
     )
     denominator = (pos_weight - neg_weight) * y_sum + 2 * neg_weight
-    return (numerator + EPS) / (denominator + EPS)
+    return numerator / (denominator + EPS)
 
 
 def jaccard_index(y_true: tf.Tensor, y_pred: tf.Tensor) -> tf.Tensor:
@@ -222,7 +224,7 @@ def jaccard_index(y_true: tf.Tensor, y_pred: tf.Tensor) -> tf.Tensor:
         + tf.reduce_sum(y_pred, axis=[1, 2, 3])
         - numerator
     )
-    return (numerator + EPS) / (denominator + EPS)
+    return numerator / (denominator + EPS)
 
 
 def gauss_kernel1d(sigma: int) -> tf.Tensor:
@@ -306,7 +308,7 @@ def compute_centroid(mask: tf.Tensor, grid: tf.Tensor) -> tf.Tensor:
     )  # (batch, dim1, dim2, dim3, 3)
     numerator = tf.reduce_sum(masked_grid, axis=[1, 2, 3])  # (batch, 3)
     denominator = tf.reduce_sum(bool_mask, axis=[1, 2, 3])  # (batch, 1)
-    return (numerator + EPS) / (denominator + EPS)  # (batch, 3)
+    return numerator / (denominator + EPS)  # (batch, 3)
 
 
 def compute_centroid_distance(
