### 2020-10-16 462deba
Issue #452: fix tests related to EPS changes
0/0 = 1 now and increase an error tolerance of one test
