### 2020-10-16 0ebe00c
PR #456: correct typos in multiple docs
