### 2020-10-17 #2705 @ethanwharris
@Kevin-Mattheus-Moerman I have now finished my review. The authors have responded to my breif comments on the paper and the code base / documentation is in good shape (installation works and the demos run as expected etc.) so I recommend acceptance :smiley:
