### 2020-10-17 ffb8b78
Issue #215: refactor lncc test in test for dissim_fn
