### 2020-10-19 e386c7b
Issue #215: Add channel dimension to Gaussian kernel (LNCC) + correct small bug in GMI
