### 2020-10-19 9cef33b
#Issue 215: Parametrize image_loss tests
