### 2020-10-19 82990c2
#Issue 215: forgot a bug in gmi
