### 2020-10-19 fca7935
#Issue 215: forgot a bug in lncc
