### 2020-10-19 a37906c
Issue #452: add eps to both numerator and denominator
