### 2020-10-21 f7e861a
Bump numpy from 1.18.5 to 1.19.2
Bumps [numpy](https://github.com/numpy/numpy) from 1.18.5 to 1.19.2.
- [Release notes](https://github.com/numpy/numpy/releases)
- [Changelog](https://github.com/numpy/numpy/blob/master/doc/HOWTO_RELEASE.rst.txt)
- [Commits](https://github.com/numpy/numpy/compare/v1.18.5...v1.19.2)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
