### 2020-10-21 18943ca
Issue #461: add stale config for Stale robot
