### 2020-10-21 2d6ac8a
Issue #344: make predict runable for grouped_mask_prostate_longitudinal
