### 2020-10-21 069c4cb
Issue #344: add test script for testing data/predict of grouped_mask_prostate_longitudinal
