+++ test/demo/test_demos.sh
@@ -1,0 +1,3 @@
+# grouped_mask_prostate_longitudinal
+python demos/grouped_mask_prostate_longitudinal/demo_data.py
+python demos/grouped_mask_prostate_longitudinal/demo_predict.py
