### 2020-10-22 ed8e84c
Issue #427 #447: Added JOSS Review Badge, update MEC info and add useful links
