+++ README.md
@@ -9,6 +9,7 @@
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![codecov](https://codecov.io/gh/DeepRegNet/DeepReg/branch/main/graph/badge.svg)](https://codecov.io/gh/DeepRegNet/DeepReg)
+[![status](https://joss.theoj.org/papers/7e6de472bc82a70d7618e23f618960b3/status.svg)](https://joss.theoj.org/papers/7e6de472bc82a70d7618e23f618960b3)
 
 **DeepReg is a freely available, community-supported open-source toolkit for research
 and education in medical image registration using deep learning.**
@@ -37,16 +38,12 @@ and education in medical image registration using deep learning.**
 
 Get involved, and help make DeepReg better!
 
-For guidance on making a contribution to DeepReg, see the
+For guidance on making a contribution to DeepReg, see our
 [contribution guidelines](https://deepreg.readthedocs.io/en/latest/contributing/code.html).
 
-If you identified a registration application with openly accessible data, please
-consider
+Have a registration application with openly accessible data? Consider
 [contributing a DeepReg Demo](https://deepreg.readthedocs.io/en/latest/contributing/demo.html).
 
 ## MICCAI 2020 Educational Challenge
 
-The MICCAI Educational Challenge notebook can be accessed
-[here](https://github.com/DeepRegNet/DeepReg/blob/main/docs/Intro_to_Medical_Image_Registration.ipynb)
-or run it on
-[Colab](https://colab.research.google.com/github/DeepRegNet/DeepReg/blob/main/docs/Intro_to_Medical_Image_Registration.ipynb).
+Our [MICCAI Educational Challenge](https://miccai-sb.github.io/materials.html) submission on DeepReg is an Award Winner! Check it out [here](https://github.com/DeepRegNet/DeepReg/blob/main/docs/Intro_to_Medical_Image_Registration.ipynb) - you can also [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/DeepRegNet/DeepReg/blob/main/docs/Intro_to_Medical_Image_Registration.ipynb)
