### 2020-10-22 b491040
Issue #427: reformat readme file
