### 2020-10-22 ef28a8f
Issue #166 #416: refactor tests
