### 2020-10-22 49c2d97
PR #456: remove strange # in cli.md
