### 2020-10-22 b0056ef
Issue #166: refactor tests for vis
