### 2020-10-22 14af081
Issue 471: update installation docs to recommend conda method.
