### 2020-10-23 e930a0a
Issue #471: uniform titles for diff options
