### 2020-10-23 DeepRegNet/DeepReg#479 @zhangfanmark
[JOSS Review] Code Review Summary
Hi,

I'm reviewing the DeepReg paper on JOSS (openjournals/joss-reviews/issues/2705).

I am able to install the software and successfully run through the code, including model training and testing, while there are some small bugs as pointed out in the other issue.

I looked through the code and find the code well-organized, well-commented, and well-documented.

I also read through the demo the authors present at MICAAI educational challenge, which gives a clear and successful demonstration of the code.
