### 2020-10-23 DeepRegNet/DeepReg#480 @zhangfanmark
[JOSS Review] Paper Review Summary 
I'm reviewing the DeepReg paper on JOSS (openjournals/joss-reviews/issues/2705).

The paper gives a clear high-level description of DeepReg and it is well written.

If space allowed, I would suggest briefly introducing the underlying algorithms (e.g., some of the text in your MIICCAI challenge demo). Also, "Statement of need" can improved by a more clear comparison to "the existing open-sourced registration projects".
