### 2020-10-23 DeepRegNet/DeepReg#479 @mathpluscode
Hi, @zhangfanmark thanks a lot for your review and comments. Please let us know if you experienced any bugs and issues (We are currently working on the demo issues you mentioned in #478 and will fix the demos and add a test soon, as mentioned in #344).
