### 2020-10-24 DeepRegNet/DeepReg#480 @mathpluscode
Hi, @zhangfanmark thanks for your comments.

Regarding the introduction of the algorithms, we had them in the paper originally however it needs more than 200 words while we already have ~800 words currently. As a result, we could not include them in the JOSS paper.

Regarding the existing open-sourced registration projects, we've mentioned
1. Deep learning based project like VoxelMorph. So far we are not aware of other related projects. Please feel free to suggest more.
2. Classical projects like NiftiReg, SimpleElastix and AirLab.

Your suggestion is about adding more projects or adding more details about these projects?

For classical projects, there is limited information to add as we are focusing on different methods. For VoxelMorph, we can update the text to point out some other problems like lack of tests.
