### 2020-10-24 b415b65
Issue #344: add max_epochs for train to overwrite config
