### 2020-10-24 be539f1
Issue #344: remove useless description for parser in cli tools
