### 2020-10-24 feaf3fd
Issue #344: refactor grouped_mask_prostate_longitudinal demo
