+++ demos/grouped_mask_prostate_longitudinal/demo_data.py
@@ -9,109 +9,123 @@ import h5py
 from scipy import ndimage
 from tensorflow.keras.utils import get_file
 
-PROJECT_DIR = r"demos/grouped_mask_prostate_longitudinal"
-os.chdir(PROJECT_DIR)
 
-DATA_PATH = "dataset"
-ZIP_FILE = "data"
-ORIGIN = "https://github.com/YipengHu/example-data/raw/master/longi-masks/data.zip"
+def main():
+    PROJECT_DIR = r"demos/grouped_mask_prostate_longitudinal"
+    os.chdir(PROJECT_DIR)
 
-if os.path.exists(DATA_PATH):
-    shutil.rmtree(DATA_PATH)
-os.mkdir(DATA_PATH)
+    DATA_PATH = "dataset"
+    ZIP_FILE = "data"
+    ORIGIN = "https://github.com/YipengHu/example-data/raw/master/longi-masks/data.zip"
 
-zip_file = os.path.join(DATA_PATH, ZIP_FILE + ".zip")
-get_file(os.path.abspath(zip_file), ORIGIN)
-with zipfile.ZipFile(zip_file, "r") as zf:
-    zf.extractall(DATA_PATH)
-os.remove(zip_file)
+    if os.path.exists(DATA_PATH):
+        shutil.rmtree(DATA_PATH)
+    os.mkdir(DATA_PATH)
 
-print("\nMask data downloaded: %s." % os.path.abspath(DATA_PATH))
+    zip_file = os.path.join(DATA_PATH, ZIP_FILE + ".zip")
+    get_file(os.path.abspath(zip_file), ORIGIN)
+    with zipfile.ZipFile(zip_file, "r") as zf:
+        zf.extractall(DATA_PATH)
+    os.remove(zip_file)
 
-## now read the data and convert to train/val/test
-ratio_val = 0.1
-ratio_test = 0.2
+    print("\nMask data downloaded: %s." % os.path.abspath(DATA_PATH))
 
-data_filename = os.path.join(DATA_PATH, ZIP_FILE + ".h5")
-fid_data = h5py.File(data_filename, "r")
-num_data = len(fid_data)
-ids_group, ids_ob = list(), list()
-for f in fid_data:
-    ds, ig, io = fid_data[f].name.split("-")
-    if ds == "/group":
-        ids_group.append(int(ig))
-        ids_ob.append(int(io))
-ids_group_unique = list(set(ids_group))
-num_group = len(ids_group_unique)
-num_val = int(num_group * ratio_val)
-num_test = int(num_group * ratio_test)
-num_train = num_group - num_val - num_test
+    ## now read the data and convert to train/val/test
+    ratio_val = 0.1
+    ratio_test = 0.2
 
-print("Found %d data in %d groups." % (num_data, num_group))
-print(
-    "Dividing into %d-%d-%d for train-val-test (%0.2f-%0.2f-%0.2f)..."
-    % (num_train, num_val, num_test, 1 - ratio_val - ratio_test, ratio_val, ratio_test)
-)
+    data_filename = os.path.join(DATA_PATH, ZIP_FILE + ".h5")
+    fid_data = h5py.File(data_filename, "r")
+    num_data = len(fid_data)
+    ids_group, ids_ob = list(), list()
+    for f in fid_data:
+        ds, ig, io = fid_data[f].name.split("-")
+        if ds == "/group":
+            ids_group.append(int(ig))
+            ids_ob.append(int(io))
+    ids_group_unique = list(set(ids_group))
+    num_group = len(ids_group_unique)
+    num_val = int(num_group * ratio_val)
+    num_test = int(num_group * ratio_test)
+    num_train = num_group - num_val - num_test
 
[…]
