### 2020-10-24 3206ab3
Issue #344: refactor grouped_mr_heart demo
