+++ test/demo/test_demos.sh
@@ -2,3 +2,8 @@
 python demos/grouped_mask_prostate_longitudinal/demo_data.py
 python demos/grouped_mask_prostate_longitudinal/demo_train.py
 python demos/grouped_mask_prostate_longitudinal/demo_predict.py
+
+# grouped_mr_heart
+python demos/grouped_mr_heart/demo_data.py
+python demos/grouped_mr_heart/demo_train.py
+python demos/grouped_mr_heart/demo_predict.py
