+++ demos/grouped_mr_heart/demo_predict.py
@@ -1,7 +1,36 @@
-"""
-This is a palceholder python file for calling the installed functions
-"""
+from deepreg.predict import predict
 
-print(
-    "Please use command line functions for this demo by following the instruction in readme.md"
-)
+name = "grouped_mr_heart"
+
+
+def main():
+    print(
+        "The prediction can also be launched using the following command."
+        "deepreg_predict --gpu ''"
+        f"--config_path demos/{name}/{name}.yaml "
+        f"--ckpt_path demos/{name}/dataset/pre-trained/weights-epoch500.ckpt "
+        f"--log_root demos/{name}"
+        "--log_dir logs_predict"
+        "--save_png --mode test"
+    )
+
+    log_root = f"demos/{name}"
+    log_dir = "logs_predict"
+    ckpt_path = f"{log_root}/dataset/pre-trained/weights-epoch500.ckpt"
+    config_path = f"{log_root}/{name}.yaml"
+
+    predict(
+        gpu="0",
+        gpu_allow_growth=False,
+        ckpt_path=ckpt_path,
+        mode="test",
+        batch_size=1,
+        log_root=log_root,
+        log_dir=log_dir,
+        sample_label="all",
+        config_path=config_path,
+    )
+
+
+if __name__ == "__main__":
+    main()
