### 2020-10-24 178bdd7
Issue #344: refactor paired_ct_lung demo
