+++ demos/paired_ct_lung/demo_predict.py
@@ -1,118 +1,36 @@
-import os
-
-import matplotlib.pyplot as plt
-
 from deepreg.predict import predict
 
-######## PREDICTION ########
+name = "paired_ct_lung"
 
 
-log_dir = "learn2reg_t2_paired_train_logs"
-
-log_dir_tr = r"demos/paired_ct_lung/learn2reg_t2_paired_train_logs"
-ckpt_path = os.path.join(log_dir_tr, "save", "weights-epoch500.ckpt")
-config_path = os.path.join(log_dir_tr, "config.yaml")
-
-gpu = ""
-gpu_allow_growth = False
-predict(
-    gpu=gpu,
-    gpu_allow_growth=gpu_allow_growth,
-    config_path=config_path,
-    ckpt_path=ckpt_path,
-    mode="test",
-    batch_size=1,
-    log_dir=log_dir,
-    sample_label="all",
-    save_png=True,
-)
-
-# the numerical metrics are saved in the logs directory specified
-
-
-######## VISUALISATION ########
-
-# Now lets load in a few samples from the predictions and plot them
-
-# change the following line to the path to image0 label0
-
-path_to_test = r"logs/learn2reg_t2_paired_train_logs/test"
-path_to_fixed_label = os.path.join(path_to_test, "pair_0", "label_0", "fixed_label")
-path_to_pred_fixed_label = os.path.join(
-    path_to_test, "pair_0", "label_0", "pred_fixed_label"
-)
-path_to_fixed_image = os.path.join(path_to_test, "pair_0", "fixed_image")
-path_to_pred_fixed_image = os.path.join(path_to_test, "pair_0", "pred_fixed_image")
-path_to_moving_image = os.path.join(path_to_test, "pair_0", "moving_image")
-path_to_moving_label = os.path.join(path_to_test, "pair_0", "label_0", "moving_label")
-
-# change inds_to_plot if different images need to be plotted instead
-
-inds_to_plot = [144, 145, 184, 140, 150, 180]
-sub_plot_counter = 1
-plt.rcParams["figure.figsize"] = [20, 20]
-
-for ind in inds_to_plot:
-    plt.subplot(6, 6, sub_plot_counter)
-    label = plt.imread(
-        os.path.join(path_to_fixed_label, "depth" + str(ind) + "_fixed_label.png")
+def main():
+    print(
+        "The prediction can also be launched using the following command."
+        "deepreg_predict --gpu ''"
+        f"--config_path demos/{name}/{name}.yaml "
+        f"--ckpt_path demos/{name}/dataset/pretrained/learn2reg_t2_paired_train_logs/save/weights-epoch500.ckpt "
+        f"--log_root demos/{name}"
+        "--log_dir logs_predict"
+        "--save_png --mode test"
     )
-    plt.imshow(label)
-    plt.axis("off")
-    if sub_plot_counter == 1:
-        plt.title("fixed_label")
 
-    plt.subplot(6, 6, sub_plot_counter + 1)
-    pred = plt.imread(
-        os.path.join(
-            path_to_pred_fixed_label, "depth" + str(ind) + "_pred_fixed_label.png"
-        )
+    log_root = f"demos/{name}"
+    log_dir = "logs_predict"
+    ckpt_path = f"{log_root}/dataset/pretrained/learn2reg_t2_paired_train_logs/save/weights-epoch500.ckpt"
+    config_path = f"{log_root}/{name}.yaml"
+
+    predict(
+        gpu="0",
+        gpu_allow_growth=False,
+        ckpt_path=ckpt_path,
+        mode="test",
+        batch_size=1,
+        log_root=log_root,
[…]
