+++ demos/paired_ct_lung/demo_train.py
@@ -1,19 +1,54 @@
+import argparse
+
 from deepreg.train import train
 
-######## NOW WE DO THE TRAINING ########
+name = "paired_ct_lung"
 
-gpu = ""
-gpu_allow_growth = False
-ckpt_path = ""
-log_dir = "learn2reg_t2_paired_train_logs"
-config_path = [
-    r"demos/paired_ct_lung/paired_ct_lung_train.yaml",
-    r"demos/paired_ct_lung/paired_ct_lung.yaml",
-]
-train(
-    gpu=gpu,
-    config_path=config_path,
-    gpu_allow_growth=gpu_allow_growth,
-    ckpt_path=ckpt_path,
-    log_dir=log_dir,
-)
+
+def launch_training(test: bool = True):
+    print(
+        "The training can also be launched using the following command."
+        "deepreg_train --gpu '0'"
+        f"--config_path demos/{name}/{name}.yaml "
+        f"--log_root demos/{name}"
+        "--log_dir logs_train"
+    )
+
+    log_root = f"demos/{name}"
+    log_dir = "logs_train"
+    config_path = [f"demos/{name}/{name}.yaml"]
+    if test:
+        config_path.append("config/test/demo_paired.yaml")
+
+    train(
+        gpu="0",
+        config_path=config_path,
+        gpu_allow_growth=False,
+        ckpt_path="",
+        log_root=log_root,
+        log_dir=log_dir,
+    )
+
+
+def main(args=None):
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--test",
+        help="Execute the script for test purpose",
+        dest="test",
+        action="store_true",
+    )
+    parser.add_argument(
+        "--no-test",
+        help="Execute the script for non-test purpose",
+        dest="test",
+        action="store_false",
+    )
+    parser.set_defaults(test=True)
+    args = parser.parse_args(args)
+
+    launch_training(test=args.test)
+
+
+if __name__ == "__main__":
+    main()
