+++ test/demo/test_demos.sh
@@ -1,9 +1,0 @@
-# grouped_mask_prostate_longitudinal
-python demos/grouped_mask_prostate_longitudinal/demo_data.py
-python demos/grouped_mask_prostate_longitudinal/demo_train.py
-python demos/grouped_mask_prostate_longitudinal/demo_predict.py
-
-# grouped_mr_heart
-python demos/grouped_mr_heart/demo_data.py
-python demos/grouped_mr_heart/demo_train.py
-python demos/grouped_mr_heart/demo_predict.py
