+++ demos/grouped_mr_heart/demo_data.py
@@ -111,8 +111,8 @@ def main():
 
     print("Done. \n")
 
-    # Download the pre-trained models
-    MODEL_PATH = os.path.join(DATA_PATH, "pre-trained")
+    # Download the pretrained models
+    MODEL_PATH = os.path.join(DATA_PATH, "pretrained")
     if os.path.exists(MODEL_PATH):
         shutil.rmtree(MODEL_PATH)
     os.mkdir(MODEL_PATH)
@@ -127,7 +127,7 @@ def main():
     os.remove(zip_file)
 
     print(
-        "Pre-trained model is downloaded and unzipped in %s."
+        "pretrained model is downloaded and unzipped in %s."
         % os.path.abspath(MODEL_PATH)
     )
 
