### 2020-10-24 8e3cc27
Issue #344: revert changes to data scripts to minimize changes
