+++ demos/paired_ct_lung/demo_data.py
@@ -10,27 +10,28 @@ from tensorflow.keras.utils import get_file
 project_dir = "demos/paired_ct_lung"
 data_folder_name = "dataset"
 
+main_path = os.getcwd()
+os.chdir(main_path)
 
-def download_data(main_path):
-    # DOWNLOADING AND UNZIPPING ALL FILES INTO CORRECT PATH
-    os.chdir(project_dir)
+# DOWNLOADING AND UNZIPPING ALL FILES INTO CORRECT PATH
+os.chdir(project_dir)
 
-    url = "https://zenodo.org/record/3835682/files/training.zip"
-    fname = "training.zip"
-    get_file(os.path.join(os.getcwd(), fname), url)
+url = "https://zenodo.org/record/3835682/files/training.zip"
+fname = "training.zip"
+get_file(os.path.join(os.getcwd(), fname), url)
 
-    print("The file ", fname, " has successfully been downloaded!")
+print("The file ", fname, " has successfully been downloaded!")
 
-    if not os.path.exists(os.path.join(main_path, project_dir, data_folder_name)):
-        os.mkdir(os.path.join(main_path, project_dir, data_folder_name))
+if not os.path.exists(os.path.join(main_path, project_dir, data_folder_name)):
+    os.mkdir(os.path.join(main_path, project_dir, data_folder_name))
 
-    with zipfile.ZipFile(fname, "r") as zip_ref:
-        zip_ref.extractall(data_folder_name)
+with zipfile.ZipFile(fname, "r") as zip_ref:
+    zip_ref.extractall(data_folder_name)
 
-    print("Files unzipped!")
+print("Files unzipped!")
 
-    os.remove(fname)
-    os.chdir(main_path)
+os.remove(fname)
+os.chdir(main_path)
 
 
 def moveTestCasesIntoCorrectPath(test_cases, path_to_train, path_to_test):
@@ -62,201 +63,172 @@ def moveFilesIntoCorrectPath(
             shutil.move(source, destination)
 
 
-def move_data(main_path):
-    # MOVING FILES INTO TRAIN DIRECTORY
-    path_to_data_folder = os.path.join(main_path, project_dir, data_folder_name)
-    path_to_train = os.path.join(main_path, project_dir, data_folder_name, "train")
-    # path_to_test = os.path.join(main_path, project_dir, data_folder_name, "test") # TODO Not used
-    path_to_images_and_labels = os.path.join(
-        main_path, project_dir, data_folder_name, "training"
+# MOVING FILES INTO TRAIN DIRECTORY
+path_to_data_folder = os.path.join(main_path, project_dir, data_folder_name)
+path_to_train = os.path.join(main_path, project_dir, data_folder_name, "train")
+# path_to_test = os.path.join(main_path, project_dir, data_folder_name, "test") # TODO Not used
+path_to_images_and_labels = os.path.join(
+    main_path, project_dir, data_folder_name, "training"
+)
+
+labels_fnames = os.listdir(os.path.join(path_to_images_and_labels, "lungMasks"))
+images_fnames = os.listdir(os.path.join(path_to_images_and_labels, "scans"))
+
+if not os.path.exists(path_to_train):
+    os.mkdir(path_to_train)
+    os.mkdir(os.path.join(path_to_train, "fixed_images"))
+    os.mkdir(os.path.join(path_to_train, "fixed_labels"))
+    os.mkdir(os.path.join(path_to_train, "moving_images"))
+    os.mkdir(os.path.join(path_to_train, "moving_labels"))
+
+if os.path.exists(path_to_images_and_labels):
+    moveFilesIntoCorrectPath(
+        path_to_train, images_fnames, path_to_images_and_labels, "images", "scans"
[…]
