### 2020-10-24 9532feb
Issue #344: revert changes to data script of paired_ct_lung
