+++ demos/paired_ct_lung/demo_data.py
@@ -6,23 +6,37 @@ import zipfile
 import nibabel as nib
 import numpy as np
 from tensorflow.keras.utils import get_file
+from tqdm import tqdm
 
-project_dir = "demos/paired_ct_lung"
-data_folder_name = "dataset"
+# if wget is installed remove the following line from comment
+# import wget
 
+# if already in the abc/DeepReg directory then do nothing, otherwise
+# use os.chdir(r'abc/DeepReg') before this line
 main_path = os.getcwd()
 os.chdir(main_path)
 
-# DOWNLOADING AND UNZIPPING ALL FILES INTO CORRECT PATH
+######## DOWNLOADING AND UNZIPPING ALL FILES INTO CORRECT PATH ########
+
+project_dir = r"demos/paired_ct_lung"
+data_folder_name = "dataset"
 os.chdir(project_dir)
 
 url = "https://zenodo.org/record/3835682/files/training.zip"
+
+# if wget is installed remove following line from comments and comment
+# out the fname = 'training.zip' line
+# fname = wget.download(url)
 fname = "training.zip"
+
+# os.system("wget " + url)
+
 get_file(os.path.join(os.getcwd(), fname), url)
 
 print("The file ", fname, " has successfully been downloaded!")
 
-if not os.path.exists(os.path.join(main_path, project_dir, data_folder_name)):
+
+if os.path.exists(os.path.join(main_path, project_dir, data_folder_name)) is not True:
     os.mkdir(os.path.join(main_path, project_dir, data_folder_name))
 
 with zipfile.ZipFile(fname, "r") as zip_ref:
@@ -33,40 +47,11 @@ print("Files unzipped!")
 os.remove(fname)
 os.chdir(main_path)
 
+######## MOVING FILES INTO TRAIN DIRECTORY ########
 
-def moveTestCasesIntoCorrectPath(test_cases, path_to_train, path_to_test):
-    folder_names = os.listdir(path_to_train)
-    os.chdir(path_to_train)
-    for case in test_cases:
-        for folder in folder_names:
-            file_names = os.listdir(os.path.join(path_to_train, folder))
-            for file in file_names:
-                if case in file:
-                    os.chdir(os.path.join(path_to_train, folder))
-                    source = file
-                    destination = os.path.join(path_to_test, folder)
-                    shutil.move(source, destination)
-
-
-def moveFilesIntoCorrectPath(
-    path_to_train, fnames, path_to_images_and_labels, suffix, sub_folder_name
-):
-    os.chdir(os.path.join(path_to_images_and_labels, sub_folder_name))
-    for file in fnames:
-        if "insp" in file:
-            source = file
-            destination = os.path.join(path_to_train, "fixed_" + suffix)
-            shutil.move(source, destination)
-        elif "exp" in file:
-            source = file
-            destination = os.path.join(path_to_train, "moving_" + suffix)
-            shutil.move(source, destination)
-
-
-# MOVING FILES INTO TRAIN DIRECTORY
 path_to_data_folder = os.path.join(main_path, project_dir, data_folder_name)
 path_to_train = os.path.join(main_path, project_dir, data_folder_name, "train")
-# path_to_test = os.path.join(main_path, project_dir, data_folder_name, "test") # TODO Not used
+path_to_test = os.path.join(main_path, project_dir, data_folder_name, "test")
 path_to_images_and_labels = os.path.join(
     main_path, project_dir, data_folder_name, "training"
 )
@@ -74,33 +59,45 @@ path_to_images_and_labels = os.path.join(
 labels_fnames = os.listdir(os.path.join(path_to_images_and_labels, "lungMasks"))
[…]
