### 2020-10-24 158560b
Issue #344: refactor paired_mrus_brain demo
