+++ demos/paired_mrus_brain/demo_data.py
@@ -1,216 +1,231 @@
+import logging
 import os
+import shutil
 import zipfile
 
+import nibabel as nib
+import numpy as np
 from tensorflow.keras.utils import get_file
+from tqdm import tqdm
 
+DOWNLOAD_FULL_DATA = False
 main_path = os.getcwd()
 
 project_dir = os.path.join(main_path, r"demos/paired_mrus_brain")
 os.chdir(project_dir)
 
-######## FULL DATA DOWNLOAD AND PREPROCESS ########
-# # Please uncomment this code block if full data needs to be used
-# from tensorflow.keras.utils import get_file
-# from tqdm import tqdm
-# import shutil
-# import nibabel as nib
-# import numpy as np
-# if os.path.exists("dataset_resect") is not True:
-#     os.mkdir("dataset_resect")
-#     os.mkdir(r"dataset_resect/paired_mr_us_brain")
-# url = "https://ns9999k.webs.sigma2.no/10.11582_2020.00025/EASY-RESECT.zip"
-# fname = "EASY-RESECT.zip"
-# path_to_zip_file = r"dataset_resect"
-# get_file(os.path.join(os.getcwd(), path_to_zip_file, fname), url)
-# with zipfile.ZipFile(os.path.join(path_to_zip_file, fname), "r") as zip_ref:
-#     zip_ref.extractall(os.path.join(path_to_zip_file, "paired_mr_us_brain"))
-# path_to_nifti = os.path.join(
-#     path_to_zip_file, "paired_mr_us_brain", "EASY-RESECT", "NIFTI"
-# )
-# all_folders = os.listdir(path_to_nifti)
-# for folder in all_folders:
-#     source = os.path.join(path_to_nifti, folder)
-#     destination = r"dataset_resect/paired_mr_us_brain"
-#     shutil.move(source, destination)
-# print("Files restructured!")
-# test_ratio = 0.25
-# path_to_data = r"dataset_resect/paired_mr_us_brain"
-# cases_list = os.listdir(path_to_data)
-# os.mkdir(os.path.join(path_to_data, "test"))
-# os.mkdir(os.path.join(path_to_data, "train"))
-# num_test = round(len(cases_list) * test_ratio)
-# for folder in cases_list[:num_test]:
-#     source = os.path.join(path_to_data, folder)
-#     destination = os.path.join(path_to_data, "test")
-#     shutil.move(source, destination)
-# for folder in cases_list[num_test:]:
-#     source = os.path.join(path_to_data, folder)
-#     destination = os.path.join(path_to_data, "train")
-#     shutil.move(source, destination)
-# folders = os.listdir(path_to_data)
-# for folder in folders:
-#     sub_folders = os.listdir(os.path.join(path_to_data, folder))
-#     for sub_folder in tqdm(sub_folders):
-#         if "DS_St" in sub_folder:
-#             os.remove(os.path.join(path_to_data, folder, sub_folder))
-#         else:
-#             files = os.listdir(os.path.join(path_to_data, folder, sub_folder))
-#             for file in files:
-#                 if "T1" in file:
-#                     arr = nib.load(
-#                         os.path.join(path_to_data, folder, sub_folder, file)
-#                     ).get_data()
-#                     img = nib.Nifti1Image(arr, affine=np.eye(4))
-#                     img.to_filename(
-#                         os.path.join(
-#                             path_to_data,
-#                             folder,
-#                             sub_folder,
-#                             file.split(".nii")[0] + "_resized.nii.gz",
-#                         )
-#                     )
-#                 elif "US" in file:
-#                     img = nib.load(os.path.join(path_to_data, folder, sub_folder, file))
-#                     nib.save(
-#                         img,
-#                         os.path.join(
-#                             path_to_data,
[…]
