+++ demos/paired_mrus_brain/demo_predict.py
@@ -1,78 +1,36 @@
-import os
-
-import matplotlib.pyplot as plt
-
 from deepreg.predict import predict
 
-######## PREDICTION ########
+name = "paired_mrus_brain"
 
-log_dir = "learn2reg_t1_paired_train_logs"
 
-log_dir_tr = "demos/paired_mrus_brain/learn2reg_t1_paired_train_logs"
-ckpt_path = os.path.join(log_dir_tr, "save", "weights-epoch800.ckpt")
-config_path = os.path.join(log_dir_tr, "config.yaml")
-
-gpu = ""
-gpu_allow_growth = False
-predict(
-    gpu=gpu,
-    gpu_allow_growth=gpu_allow_growth,
-    config_path=config_path,
-    ckpt_path=ckpt_path,
-    mode="test",
-    batch_size=1,
-    log_dir=log_dir,
-    sample_label="all",
-    save_png=True,
-)
-
-# the numerical metrics are saved in the specified logs directory
-
-######## VISUALISATION ########
-# now let's load a few samples from the predictions and plot them
-# change the following line to the path to image0 label0
-path_to_image0_label0 = r"logs/learn2reg_t1_paired_train_logs/test"
-path_to_pred_fixed_img = os.path.join(path_to_image0_label0, r"pair_0/pred_fixed_image")
-path_to_moving_img = os.path.join(path_to_image0_label0, r"pair_0/moving_image")
-path_to_fixed_img = os.path.join(path_to_image0_label0, r"pair_0/fixed_image")
-
-# change inds_to_plot if different images need to be plotted instead
-inds_to_plot = [50, 120, 150, 160, 190, 210]
-sub_plot_counter = 1
-
-for ind in inds_to_plot:
-    plt.subplot(6, 3, sub_plot_counter)
-    label = plt.imread(
-        os.path.join(path_to_moving_img, "depth" + str(ind) + "_moving_image.png")
+def main():
+    print(
+        "The prediction can also be launched using the following command."
+        "deepreg_predict --gpu ''"
+        f"--config_path demos/{name}/{name}.yaml "
+        f"--ckpt_path demos/{name}/dataset/pretrained/learn2reg_t1_paired_train_logs/save/weights-epoch800.ckpt "
+        f"--log_root demos/{name}"
+        "--log_dir logs_predict"
+        "--save_png --mode test"
     )
-    plt.imshow(label)
-    plt.axis("off")
-    if sub_plot_counter == 1:
-        plt.title("Moving Image")
 
-    plt.subplot(6, 3, sub_plot_counter + 1)
-    pred = plt.imread(
-        os.path.join(
-            path_to_pred_fixed_img, "depth" + str(ind) + "_pred_fixed_image.png"
-        )
+    log_root = f"demos/{name}"
+    log_dir = "logs_predict"
+    ckpt_path = f"{log_root}/dataset/pretrained/learn2reg_t1_paired_train_logs/save/weights-epoch800.ckpt"
+    config_path = f"{log_root}/{name}.yaml"
+
+    predict(
+        gpu="0",
+        gpu_allow_growth=False,
+        ckpt_path=ckpt_path,
+        mode="test",
+        batch_size=1,
+        log_root=log_root,
+        log_dir=log_dir,
+        sample_label="all",
+        config_path=config_path,
     )
-    plt.imshow(pred)
-    plt.axis("off")
-    if sub_plot_counter == 1:
-        plt.title("Warped Moving Image")
 
-    plt.subplot(6, 3, sub_plot_counter + 2)
-    pred = plt.imread(
-        os.path.join(path_to_fixed_img, "depth" + str(ind) + "_fixed_image.png")
-    )
[…]
