### 2020-10-24 ae93f8b
Issue #344: refactor paired_mrus_prostate demo
