+++ demos/unpaired_ct_lung/README.md
@@ -27,7 +27,7 @@ application can be seen in [2].
   the following command:
 
 ```bash
-python3 demos/unpaired_ct_lung/script_name.py
+python demos/unpaired_ct_lung/script_name.py
 ```
 
 A short description of the scripts is provided below. The scripts must be run in the

[…]
