### 2020-10-25 89f5e8b
Issue #344: refactor unpaired_ct_abdomen demo
