+++ test/demo/test_demos.py
@@ -5,17 +5,7 @@ import subprocess
 import pytest
 
 
-@pytest.mark.parametrize(
-    "name",
-    [
-        "grouped_mask_prostate_longitudinal",
-        "grouped_mr_heart",
-        "paired_ct_lung",
-        "paired_mrus_brain",
-        "paired_mrus_prostate",
-    ],
-)
-def test_demo(name):
+def remove_files(name):
     dir_name = os.path.join("demos", name)
 
     # remove zip files
@@ -32,9 +22,8 @@ def test_demo(name):
         if os.path.exists(path):
             shutil.rmtree(path)
 
-    # execute data, train, predict sequentially
-    cmds = [f"python demos/{name}/demo_{x}.py" for x in ["data", "train", "predict"]]
 
+def execute_commands(cmds):
     for cmd in cmds:
         try:
             print(f"Running {cmd}")
@@ -44,3 +33,41 @@ def test_demo(name):
             raise RuntimeError(
                 f"Command {cmd} return with err {e.returncode} {e.output}"
             )
+
+
+class TestDemo:
+    @pytest.mark.parametrize(
+        "name",
+        [
+            "grouped_mask_prostate_longitudinal",
+            "grouped_mr_heart",
+            "paired_ct_lung",
+            "paired_mrus_brain",
+            "paired_mrus_prostate",
+        ],
+    )
+    def test_simple_demo(self, name):
+        """each demo has one single configuration file"""
+        remove_files(name)
+
+        # execute data, train, predict sequentially
+        cmds = [
+            f"python demos/{name}/demo_{x}.py" for x in ["data", "train", "predict"]
+        ]
+
+        execute_commands(cmds)
+
+    def test_unpaired_ct_abdomen(self):
+        """this demo has multiple configuration file"""
+        name = "unpaired_ct_abdomen"
+        remove_files(name)
+
+        # execute data, train, predict sequentially
+        cmds = [f"python demos/{name}/demo_data.py"]
+        for method in ["comb", "unsup", "weakly"]:
+            cmds += [
+                f"python demos/{name}/demo_{x}.py --method {method}"
+                for x in ["train", "predict"]
+            ]
+
+        execute_commands(cmds)
