### 2020-10-25 a56782f
Issue #344: refactor unpaired_ct_lung demo
