+++ demos/unpaired_ct_abdomen/demo_train.py
@@ -8,9 +8,9 @@ name = "unpaired_ct_abdomen"
 def launch_training(method: str, test: bool = True):
     print(
         "The training can also be launched using the following command."
-        "deepreg_train --gpu '0'"
+        "deepreg_train --gpu '0' "
         f"--config_path demos/{name}/{name}_{method}.yaml "
-        f"--log_root demos/{name}"
+        f"--log_root demos/{name} "
         f"--log_dir logs_train/{method}"
     )
 

[…]
