### 2020-10-25 ebc1047
Issue #344: refactor unpaired_mr_brain demo
