+++ demos/unpaired_mr_brain/demo_predict.py
@@ -1,123 +1,36 @@
-import os
-
-import matplotlib.pyplot as plt
-
 from deepreg.predict import predict
 
-######## PREDICTION ########
-# saved weights: change it if you retrained the model.
-log_dir = r"learn2reg_t4_unpaired_weakly_train_logs2"
-ckpt_path = os.path.join("logs", log_dir, "save", "weights-epoch200.ckpt")
-config_path = [os.path.join("logs", log_dir, "config.yaml")]
-gpu = "0"
-gpu_allow_growth = False
-predict(
-    gpu=gpu,
-    gpu_allow_growth=gpu_allow_growth,
-    ckpt_path=ckpt_path,
-    mode="test",
-    batch_size=1,
-    log_dir=log_dir,
-    sample_label="all",
-    config_path=config_path,
-)
-
-# the numerical metrics are saved in the logs directory specified
+name = "unpaired_mr_brain"
 
 
-######## VISUALISATION ########
-
-# Now lets load in a few samples from the predictions and plot them
-main_path = os.getcwd()
-os.chdir(main_path)
-
-# change the following line to the path to image0 label0
-path_to_image0_label0 = os.path.join("logs", log_dir, "test")
-path_to_image0_label0 = os.path.join(main_path, path_to_image0_label0)
-
-path_to_pred_fixed_label = os.path.join(
-    path_to_image0_label0, r"pair_0_1/label_0/pred_fixed_label"
-)
-path_to_fixed_label = os.path.join(
-    path_to_image0_label0, r"pair_0_1/label_0/fixed_label"
-)
-
-path_to_fixed_image = os.path.join(path_to_image0_label0, "pair_0_1", "fixed_image")
-path_to_pred_fixed_image = os.path.join(
-    path_to_image0_label0, "pair_0_1", "pred_fixed_image"
-)
-path_to_moving_image = os.path.join(path_to_image0_label0, "pair_0_1", "moving_image")
-path_to_moving_label = os.path.join(
-    path_to_image0_label0, "pair_0_1", "label_0", "moving_label"
-)
-
-# change inds_to_plot if different images need to be plotted instead
-
-inds_to_plot = [10, 15, 20, 25, 30, 35]
-sub_plot_counter = 1
-plt.rcParams["figure.figsize"] = [20, 20]
-
-for ind in inds_to_plot:
-    plt.subplot(6, 6, sub_plot_counter)
-    label = plt.imread(
-        os.path.join(path_to_fixed_label, "depth" + str(ind) + "_fixed_label.png")
+def main():
+    print(
+        "The prediction can also be launched using the following command."
+        "deepreg_predict --gpu ''"
+        f"--config_path demos/{name}/{name}.yaml "
+        f"--ckpt_path demos/{name}/dataset/pretrained/learn2reg_t4_unpaired_weakly_train_logs2/save/weights-epoch200.ckpt "
+        f"--log_root demos/{name} "
+        "--log_dir logs_predict"
+        "--save_png --mode test"
     )
-    plt.imshow(label)
-    plt.axis("off")
-    if sub_plot_counter == 1:
-        plt.title("fixed_label")
 
-    plt.subplot(6, 6, sub_plot_counter + 1)
-    pred = plt.imread(
-        os.path.join(
-            path_to_pred_fixed_label, "depth" + str(ind) + "_pred_fixed_label.png"
-        )
+    log_root = f"demos/{name}"
+    log_dir = "logs_predict"
+    ckpt_path = f"{log_root}/dataset/pretrained/learn2reg_t4_unpaired_weakly_train_logs2/save/weights-epoch200.ckpt"
[…]
