+++ demos/unpaired_mr_brain/demo_train.py
@@ -1,21 +1,54 @@
-import os
+import argparse
 
 from deepreg.train import train
 
-os.environ["CUDA_VISIBLE_DEVICES"] = "0"
+name = "unpaired_mr_brain"
 
-gpu = "0"
-gpu_allow_growth = False
-ckpt_path = ""  # To load pre-trained weights
-config_path = [r"demos/unpaired_mr_brain/unpaired_mr_brain.yaml"]
 
-# log_dir: this log dir points to the downloaded log dir. Change it for other experiments.
-log_dir = "learn2reg_t4_unpaired_weakly_subset_train_logs"
+def launch_training(test: bool = True):
+    print(
+        "The training can also be launched using the following command."
+        "deepreg_train --gpu '0' "
+        f"--config_path demos/{name}/{name}.yaml "
+        f"--log_root demos/{name} "
+        "--log_dir logs_train"
+    )
 
-train(
-    gpu=gpu,
-    config_path=config_path,
-    gpu_allow_growth=gpu_allow_growth,
-    ckpt_path=ckpt_path,
-    log_dir=log_dir,
-)
+    log_root = f"demos/{name}"
+    log_dir = "logs_train"
+    config_path = [f"demos/{name}/{name}.yaml"]
+    if test:
+        config_path.append("config/test/demo_unpaired_grouped.yaml")
+
+    train(
+        gpu="0",
+        config_path=config_path,
+        gpu_allow_growth=False,
+        ckpt_path="",
+        log_root=log_root,
+        log_dir=log_dir,
+    )
+
+
+def main(args=None):
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--test",
+        help="Execute the script for test purpose",
+        dest="test",
+        action="store_true",
+    )
+    parser.add_argument(
+        "--no-test",
+        help="Execute the script for non-test purpose",
+        dest="test",
+        action="store_false",
+    )
+    parser.set_defaults(test=True)
+    args = parser.parse_args(args)
+
+    launch_training(test=args.test)
+
+
+if __name__ == "__main__":
+    main()
