### 2020-10-25 06dce8d
Issue #344: refactor unpaired_us_prostate_cv demo
