### 2020-10-25 e747e37
Issue #344: refactor classical demos
