+++ demos/classical_mr_prostate_nonrigid/demo_register.py
@@ -1,8 +1,8 @@
 """
 A DeepReg Demo for classical nonrigid iterative pairwise registration algorithms
 """
+import argparse
 import os
-from datetime import datetime
 
 import h5py
 import tensorflow as tf
@@ -12,118 +12,146 @@ import deepreg.model.loss.deform as deform_loss
 import deepreg.model.loss.image as image_loss
 import deepreg.util as util
 
-MAIN_PATH = os.getcwd()
-PROJECT_DIR = "demos/classical_mr_prostate_nonrigid"
-os.chdir(PROJECT_DIR)
 
-DATA_PATH = "dataset"
-FILE_PATH = os.path.join(DATA_PATH, "demo2.h5")
+def launch_registration(test: bool):
+    MAIN_PATH = os.getcwd()
+    PROJECT_DIR = "demos/classical_mr_prostate_nonrigid"
+    os.chdir(PROJECT_DIR)
 
-# registration parameters
-image_loss_name = "lncc"
-deform_loss_name = "bending"
-weight_deform_loss = 1
-learning_rate = 0.1
-total_iter = int(3000)
+    DATA_PATH = "dataset"
+    FILE_PATH = os.path.join(DATA_PATH, "demo2.h5")
 
-# load images
-if not os.path.exists(DATA_PATH):
-    raise ("Download the data using demo_data.py script")
-if not os.path.exists(FILE_PATH):
-    raise ("Download the data using demo_data.py script")
+    # registration parameters
+    image_loss_name = "lncc"
+    deform_loss_name = "bending"
+    weight_deform_loss = 1
+    learning_rate = 0.1
+    total_iter = int(10) if test else int(3000)
 
-fid = h5py.File(FILE_PATH, "r")
-moving_image = tf.cast(tf.expand_dims(fid["image0"], axis=0), dtype=tf.float32)
-fixed_image = tf.cast(tf.expand_dims(fid["image1"], axis=0), dtype=tf.float32)
+    # load images
+    if not os.path.exists(DATA_PATH):
+        raise ("Download the data using demo_data.py script")
+    if not os.path.exists(FILE_PATH):
+        raise ("Download the data using demo_data.py script")
 
+    fid = h5py.File(FILE_PATH, "r")
+    moving_image = tf.cast(tf.expand_dims(fid["image0"], axis=0), dtype=tf.float32)
+    fixed_image = tf.cast(tf.expand_dims(fid["image1"], axis=0), dtype=tf.float32)
 
-# optimisation
-@tf.function
-def train_step(warper, weights, optimizer, mov, fix):
-    """
-    Train step function for backprop using gradient tape
+    # optimisation
+    @tf.function
+    def train_step(warper, weights, optimizer, mov, fix):
+        """
+        Train step function for backprop using gradient tape
 
-    :param warper: warping function returned from layer.Warping
-    :param weights: trainable ddf [1, f_dim1, f_dim2, f_dim3, 3]
-    :param optimizer: tf.optimizers
-    :param mov: moving image [1, m_dim1, m_dim2, m_dim3]
-    :param fix: fixed image [1, f_dim1, f_dim2, f_dim3]
-    :return:
-        loss: overall loss to optimise
-        loss_image: image dissimilarity
-        loss_deform: deformation regularisation
-    """
-    with tf.GradientTape() as tape:
-        pred = warper(inputs=[weights, mov])
-        loss_image = image_loss.dissimilarity_fn(
-            y_true=fix, y_pred=pred, name=image_loss_name
-        )
-        loss_deform = deform_loss.local_displacement_energy(weights, deform_loss_name)
-        loss = loss_image + weight_deform_loss * loss_deform
[…]
