+++ demos/classical_ct_headneck_affine/demo_register.py
@@ -1,8 +1,8 @@
 """
 A DeepReg Demo for classical affine iterative pairwise registration algorithms
 """
+import argparse
 import os
-from datetime import datetime
 
 import h5py
 import tensorflow as tf
@@ -11,126 +11,152 @@ import deepreg.model.layer_util as layer_util
 import deepreg.model.loss.image as image_loss
 import deepreg.util as util
 
-MAIN_PATH = os.getcwd()
-PROJECT_DIR = "demos/classical_ct_headneck_affine"
-os.chdir(PROJECT_DIR)
 
-DATA_PATH = "dataset"
-FILE_PATH = os.path.join(DATA_PATH, "demo.h5")
+def launch_registration(test: bool):
+    MAIN_PATH = os.getcwd()
+    PROJECT_DIR = "demos/classical_ct_headneck_affine"
+    os.chdir(PROJECT_DIR)
 
-## registration parameters
-image_loss_name = "ssd"
-learning_rate = 0.01
-total_iter = int(1000)
+    DATA_PATH = "dataset"
+    FILE_PATH = os.path.join(DATA_PATH, "demo.h5")
 
-## load image
-if not os.path.exists(DATA_PATH):
-    raise ("Download the data using demo_data.py script")
-if not os.path.exists(FILE_PATH):
-    raise ("Download the data using demo_data.py script")
+    ## registration parameters
+    image_loss_name = "ssd"
+    learning_rate = 0.01
+    total_iter = int(10) if test else int(1000)
 
-fid = h5py.File(FILE_PATH, "r")
-fixed_image = tf.cast(tf.expand_dims(fid["image"], axis=0), dtype=tf.float32)
-fixed_image = (fixed_image - tf.reduce_min(fixed_image)) / (
-    tf.reduce_max(fixed_image) - tf.reduce_min(fixed_image)
-)  # normalisation to [0,1]
+    ## load image
+    if not os.path.exists(DATA_PATH):
+        raise ("Download the data using demo_data.py script")
+    if not os.path.exists(FILE_PATH):
+        raise ("Download the data using demo_data.py script")
 
-# generate a radomly-affine-transformed moving image
-fixed_image_size = fixed_image.shape
-transform_random = layer_util.random_transform_generator(batch_size=1, scale=0.2)
-grid_ref = layer_util.get_reference_grid(grid_size=fixed_image_size[1:4])
-grid_random = layer_util.warp_grid(grid_ref, transform_random)
-moving_image = layer_util.resample(vol=fixed_image, loc=grid_random)
-# warp the labels to get ground-truth using the same random affine, for validation
-fixed_labels = tf.cast(tf.expand_dims(fid["label"], axis=0), dtype=tf.float32)
-moving_labels = tf.stack(
-    [
-        layer_util.resample(vol=fixed_labels[..., idx], loc=grid_random)
-        for idx in range(fixed_labels.shape[4])
-    ],
-    axis=4,
-)
+    fid = h5py.File(FILE_PATH, "r")
+    fixed_image = tf.cast(tf.expand_dims(fid["image"], axis=0), dtype=tf.float32)
+    fixed_image = (fixed_image - tf.reduce_min(fixed_image)) / (
+        tf.reduce_max(fixed_image) - tf.reduce_min(fixed_image)
+    )  # normalisation to [0,1]
 
+    # generate a radomly-affine-transformed moving image
+    fixed_image_size = fixed_image.shape
+    transform_random = layer_util.random_transform_generator(batch_size=1, scale=0.2)
+    grid_ref = layer_util.get_reference_grid(grid_size=fixed_image_size[1:4])
+    grid_random = layer_util.warp_grid(grid_ref, transform_random)
[…]
