### 2020-10-25 a3b9f85
Issue #344: update demo requirements and readme template
