+++ docs/source/contributing/demo.md
@@ -20,7 +20,8 @@ DeepReg Demo.
 
 - Each demo _must_ have a `demo_data.py` script to automatically download and preprocess
   demo data;
-- Data should be downloaded under the demo folder named `dataset`;
+- Data for training and test should be downloaded under the demo folder named `dataset`,
+  such as `dataset/train` and `dataset/test`;
 - Data should be hosted in a reliable and efficient (DeepReg repo will not store demo
   data or model) online storage, Kaggle, GitHub and Zendoo are all options for non-login
   access (avoid google drive for known accessibility issues);
@@ -35,34 +36,22 @@ DeepReg Demo.
   github.com/DeepRegNet/deepreg-model-zoo being preferred for storing the models. Please
   contact the Development Team for access;
 - The pre-trained model, e.g. ckpt files, should be downloaded and extracted under the
-  `dataset` folder. Avoid overwriting with user-trained models;
+  `dataset/pretrained` folder. Avoid overwriting with user-trained models;
 
 ## Training
 
-- Each demo _must_ have a `demo_train.py` script. If using command line interface
-  `deepreg_train`, this file needs to print a message to direct the user to the
-  readme.md file (described below) for instructions;
-- This is accompanied by one or more config yaml files in the same folder. If
-  appropriate, please use the same demo folder name for the config file. Add postfix if
-  using multiple config files, e.g. `unpaired_lung_ct_dataset.yaml`,
-  `unpaired_lung_ct_train.yaml`.
+- Each demo _must_ have a `demo_train.py` script;
+- This is accompanied by one or more config yaml files in the same folder. Please use
+  the same demo folder name for the config file. Add postfix if multiple training
+  methods are provided, e.g. `unpaired_ct_abdomen_comb.yaml`,
+  `unpaired_ct_abdomen_unsup.yaml`.
 
 ## Predicting
 
-- Each demo _must_ have a `demo_predict.py` script; If using command line interface
-  `deepreg_predict`, this file needs to print a message to direct users to the readme.md
-  file (described below) for instructions;
+- Each demo _must_ have a `demo_predict.py` script;
 - By default, the pre-trained model should be used in `demo_predict.py`. However, the
   instruction should be clearly given to use the user-trained model, saved with the
   `demo_train.py`;
-- Report registration results. Provide at least one piece of numerical metric (Dice,
-  distance error, etc) to show the efficacy of the registration. Optimum performance is
-  not required;
-- Provide at least one piece of visualisation of the results, such as moving image vs
-  fixed image vs warped moving image (pred_fixed_image). This may be simply done by
-  selecting the typical results from the predict output. If possible, save the
-  visualisation to (e.g. png/jpg) files, avoiding compatibility issues. Pointing to the
-  relevant file paths generated using `deepreg_predict` is recommended.
 
 ## A README.md file
 
@@ -71,11 +60,10 @@ based on the [template](../demo/readme_template.html).
 
 Following is a checklist for modifying the README template:
 
-- Modify the link to source code;
-- Modify the author section;
-- Modify the application section;
-- Modify the data section;
-- Modify the steps in instruction section;
-- Modify the pre-trained model section;
-- Modify the tested version;
-- Modify the reference section.
+- [ ] Update the link to source code;
+- [ ] Update the author section;
+- [ ] Update the application section;
+- [ ] Update the data section;
+- [ ] Update the `name` in all commands;
+- [ ] Update the instruction section for extra instructions;
+- [ ] Update the reference section.
