### 2020-10-25 17e3aae
Issue #344: update classical demo readme
