### 2020-10-25 33f3488
Issue #344: update grouped demo readme
