+++ demos/grouped_mr_heart/README.md
@@ -20,54 +20,66 @@ the late gadolinium enhancement (LGE) CMR sequence MI, the T2-weighted CMR. They
 collectively provide radiological information otherwise unavailable during clinical
 practice.
 
-## Instruction
-
-- [Install DeepReg](https://deepreg.readthedocs.io/en/latest/getting_started/install.html);
-- Change current directory to the root directory of DeepReg project;
-- Run `demo_data.py` script to download all the CMR dataset in a zip file. The script
-  also splits the data into train, val and test sets re-samples all the images to an
-  isotropic voxel size.
-
-```bash
-python demos/grouped_mr_heart/demo_data.py
-```
-
-- Run `demo_train` script to launch the training. The following example uses a single
-  GPU and launches the first of the ten runs of a 9-fold cross-validation, as specified
-  in the [`dataset` section](./grouped_mr_heart.yaml) and the
-  [`train` section](./grouped_mr_heart.yaml), which can be specified in
-  [seperate yaml files](https://deepregnet.github.io/DeepReg/#/tutorial_experiment?id=cross-validation).
-  The logs will be saved under `logs_train/` inside the demo folder;
-
-```bash
-python demos/grouped_mr_heart/demo_train.py --no-test
-```
-
-- Run `demo_predict` script to use the saved checkpoint file for testing on the data
-  partitions specified in the config file, a copy of which will be saved in the log_dir.
-  The following example uses a pre-trained model, on CPU. The results will be saved at
-  under `logs_predict/` inside the demo folder.
-
-```bash
-python demos/grouped_mr_heart/demo_predict.py
-```
-
-## Pre-trained Model
-
-A pre-trained model will be downloaded after running `demo_data.py` and unzipped at
-dataset folder under the demo folder. This pre-trained model will be used by default
-with `deepreg_predict`. Run the user-trained model by specifying with `--ckpt_path` the
-location where the checkpoint files are saved.
-
 ## Data
 
 This demo uses CMR images from 45 patients, acquired from the
 [MyoPS2020](http://www.sdspeople.fudan.edu.cn/zhuangxiahai/0/MyoPS20/) challenge held in
 conjunction with MICCAI 2020.
 
+## Instruction
+
+### Installation
+
+Please install DeepReg following the [instructions](../getting_started/install.html) and
+change the current directory to the root directory of DeepReg project, i.e. `DeepReg/`.
+
+### Download data
+
+Please execute the following command to download the data and pre-trained model. Images
+are re-sampled to an isotropic voxel size.
+
+```bash
+python demos/grouped_mr_heart/demo_data.py
+```
+
+### Launch demo training
+
+Please execute the following command to launch a demo training (the first of the ten
+runs of a 9-fold cross-validation). The training logs will be saved under
+`demos/grouped_mr_heart/logs_train`, where the saved checkpoints can be used for
+prediction later.
+
+```bash
+python demos/grouped_mr_heart/demo_train.py
+```
+
+Here the training is launched using the GPU of index 0 with a limited number of steps
+and reduced size. Please add flag `--no-test` to use the original training
+configuration, such as
+
+```bash
+python demos/grouped_mr_heart/demo_train.py --no-test
+```
+
+### Launch prediction
+
+Please execute the following command to launch the prediction with pre-trained model.
+The prediction logs will be saved under `demos/grouped_mr_heart/logs_predict`, where the
+visualization of predictions are saved. Check the [CLI documentation](../docs/cli.html)
+for more details about prediction output.
+
+```bash
+python demos/grouped_mr_heart/demo_predict.py
+```
+
+Optionally, the user-trained model can be used by changing the `ckpt_path` variable
[…]
