### 2020-10-25 68daa4f
Issue #344: update paired demo readme
