### 2020-10-25 30c0725
Issue #344: update unpaired demo readme
