+++ demos/unpaired_us_prostate_cv/README.md
@@ -25,47 +25,54 @@ dataset, hosted at the
 
 ## Instruction
 
-<!---
-"""bash config_generator cross --data_folders dataset/fold0 dataset/fold1 dataset/fold2
-dataset/fold3 dataset/fold4 dataset/fold5 dataset/fold6 dataset/fold7 dataset/fold8
-dataset/fold9 --prefix unpaired_us_prostate_cv_run
--->
+### Installation
 
-- [Install DeepReg](https://deepreg.readthedocs.io/en/latest/getting_started/install.html);
-- Change current directory to the root directory of DeepReg project;
-- Run `demo_data.py` script to download 10 folds of unpaired 3D ultrasound images;
+Please install DeepReg following the [instructions](../getting_started/install.html) and
+change the current directory to the root directory of DeepReg project, i.e. `DeepReg/`.
+
+### Download data
+
+Please execute the following command to download and pre-process the data and
+pre-trained model. Data are split into 10 folds for cross-validation.
 
 ```bash
 python demos/unpaired_us_prostate_cv/demo_data.py
 ```
 
-- Call `deepreg_train` from command line. The following example uses three GPUs and
-  launches the first of the ten runs of a 9-fold cross-validation, as specified in the
-  [`dataset` section](./unpaired_us_prostate_cv_run1.yaml) and the
-  [`train` section](./unpaired_us_prostate_cv_train.yaml), which can be specified in
-  [separate yaml files](https://deepreg.readthedocs.io/en/latest/tutorial/cross_val.html).
-  The 10th fold is reserved for testing;
+### Launch demo training
+
+Please execute the following command to launch a demo training (the first of the ten
+runs of a 9-fold cross-validation). The training logs and model checkpoints will be
+saved under `demos/unpaired_us_prostate_cv/logs_train`.
+
+```bash
+python demos/unpaired_us_prostate_cv/demo_train.py
+```
+
+Here the training is launched using the GPU of index 0 with a limited number of steps
+and reduced size. Please add flag `--no-test` to use the original training
+configuration, such as
 
 ```bash
 python demos/unpaired_us_prostate_cv/demo_train.py --no-test
 ```
 
-- Call `deepreg_predict` from command line to use the saved ckpt file for testing on the
-  10th fold data. The following example uses a pre-trained model, on CPU. If not
-  specified, the results will be saved at the created timestamp-named directories under
-  `logs_predict/` inside the demo folder.
+### Launch prediction
+
+Please execute the following command to launch the prediction with pre-trained model.
+The prediction logs and visualization results will be saved under
+`demos/unpaired_us_prostate_cv/logs_predict`. Check the
+[CLI documentation](../docs/cli.html) for more details about prediction output.
 
 ```bash
 python demos/unpaired_us_prostate_cv/demo_predict.py
 ```
 
-## Pre-trained Model
-
-A pre-trained model will be downloaded after running `demo_data.py` and unzipped at the
-dataset folder under the demo folder. This pre-trained model will be used by default
-with `deepreg_predict`. Run the user-trained model by specifying with `--ckpt_path` the
-location where the checkpoint files are saved.
+Optionally, the user-trained model can be used by changing the `ckpt_path` variable
+inside `demo_predict.py`. Note that the path should end with `.ckpt` and checkpoints are
+saved under `logs_train` as mentioned above.
 
 ## Contact
 
-Please [raise an issue](https://github.com/DeepRegNet/DeepReg/issues/new/choose).
+Please [raise an issue](https://github.com/DeepRegNet/DeepReg/issues/new/choose) for any
+questions.
