+++ demos/classical_mr_prostate_nonrigid/README.md
@@ -33,7 +33,7 @@ change the current directory to the root directory of DeepReg project, i.e. `Dee
 
 ### Download data
 
-Please execute the following command to download the data.
+Please execute the following command to download and pre-process the data.
 
 ```bash
 python demos/classical_mr_prostate_nonrigid/demo_data.py
