+++ docs/source/demo/readme_template.md
@@ -27,7 +27,8 @@ change the current directory to the root directory of DeepReg project, i.e. `Dee
 
 ### Download data
 
-Please execute the following command to download the data and pre-trained model.
+Please execute the following command to download and pre-process the data and
+pre-trained model.
 
 ```bash
 python demos/name/demo_data.py
@@ -35,9 +36,8 @@ python demos/name/demo_data.py
 
 ### Launch demo training
 
-Please execute the following command to launch a demo training. The training logs will
-be saved under `demos/name/logs_train`, where the saved checkpoints can be used for
-prediction later.
+Please execute the following command to launch a demo training. The training logs and
+model checkpoints will be saved under `demos/name/logs_train`.
 
 ```bash
 python demos/name/demo_train.py
@@ -54,9 +54,9 @@ python demos/name/demo_train.py --no-test
 ### Launch prediction
 
 Please execute the following command to launch the prediction with pre-trained model.
-The prediction logs will be saved under `demos/name/logs_predict`, where the
-visualization of predictions are saved. Check the [CLI documentation](../docs/cli.html)
-for more details about prediction output.
+The prediction logs and visualization results will be saved under
+`demos/name/logs_predict`. Check the [CLI documentation](../docs/cli.html) for more
+details about prediction output.
 
 ```bash
 python demos/name/demo_predict.py

[…]
