### 2020-10-25 a2f578d
Issue #344: update changelog
