+++ CHANGELOG.md
@@ -14,8 +14,10 @@ straightforward as possible.
 
 - Added option to change kernel size and a kernel type for LNCC image similarity loss.
 - Added visualization tool for generating gifs from model outputs.
+- Added max_epochs argument for training to overwrite configuration.
 - Added log_root argument for training and prediction to customize the log file
   location.
+- Added tests for all demos.
 - Added environment.yml file for Conda environment creation.
 - Added Dockerfile.
 - Added documentation about using UCL cluster with DeepReg.
@@ -24,10 +26,11 @@ straightforward as possible.
 
 - Updated instruction on Conda environment creation.
 - Updated TensorFlow version to 2.3.1.
-- Updated the issue and pull-request templates to fix some
 - Updated the pre-trained models in MR brain demo.
+- Updated the instructions for all demos.
+- Updated the issue and pull-request templates.
 - Updated pre-commit hooks version.
-- Updated references in JOSS paper to address reviewers' comments.
+- Updated JOSS paper to address reviewers' comments.
 
 ### Fixed
 
