### 2020-10-25 2c7e3bc
Issue #344: remove logs_reg before tests
