### 2020-10-26 DeepRegNet/DeepReg#483 @lbrea
[JOSS Review] Classical registration demos 
Hi, I'm reviewing the JOSS submission (https://github.com/openjournals/joss-reviews/issues/2705)

I've successfully run all the demos, but on the two examples of classical registration,

`python3 demos/classical_mr_prostate_nonrigid/demo_register.py`

and

`python3 demos/classical_ct_headneck_affine/demo_register.py`

the execution finishes with the following error:

`Traceback (most recent call last):`
`  File "demos/classical_mr_prostate_nonrigid/demo_register.py", line 126, in <module>`
`    save_dir=SAVE_PATH, arr=arr, name=arr_name, normalize=True, save_png=False`
`TypeError: save_array() got an unexpected keyword argument 'normalize'`
