### 2020-10-26 DeepRegNet/DeepReg#484 @lbrea
[JOSS Review] Summary
Hi, I'm reviewing the JOSS submission (https://github.com/openjournals/joss-reviews/issues/2705)

I've successfully installed DeepReg, run the demos, and check a few other examples - with only minor erros, as reported in other issue.

The documentation is very informative, and covers from installation and use to how to contribute, including many examples.
The code itself appears well structured and clear, and includes appropriate tests.

I found specially interesting how it is straighforward to include and test new components, and that it incorporates an easy way to apply cross-validation.
