### 2020-10-26 #2705 @lbrea
@Kevin-Mattheus-Moerman I've finished my review (see further comments above). The package looks great to me, and I would recommend acceptance.
