### 2020-10-26 DeepRegNet/DeepReg#483 @YipengHu
@mathpluscode is this being dealt with at #344 ?
