### 2020-10-26 DeepRegNet/DeepReg#484 @YipengHu
Nice to see the reviewer highlighted on the technical details such as the cross-validation - really impressed with the depth of the JOSS reviews!
