### 2020-10-26 DeepRegNet/DeepReg#483 @mathpluscode
> @mathpluscode is this being dealt with at #344 ?

Yes, all demos should work directly after resolving #344, as tests will be included.
