### 2020-10-27 700c1ba
Issue #344: refactor demos
- remove data folders inside data_demo
- flatten train/predict scripts
- add timestamp to log dirs
