+++ demos/unpaired_mr_brain/demo_train.py
@@ -1,54 +1,46 @@
 import argparse
+from datetime import datetime
 
 from deepreg.train import train
 
 name = "unpaired_mr_brain"
 
 
-def launch_training(test: bool = True):
-    print(
-        "The training can also be launched using the following command."
-        "deepreg_train --gpu '0' "
-        f"--config_path demos/{name}/{name}.yaml "
-        f"--log_root demos/{name} "
-        "--log_dir logs_train"
-    )
+parser = argparse.ArgumentParser()
+parser.add_argument(
+    "--test",
+    help="Execute the script for test purpose",
+    dest="test",
+    action="store_true",
+)
+parser.add_argument(
+    "--no-test",
+    help="Execute the script for non-test purpose",
+    dest="test",
+    action="store_false",
+)
+parser.set_defaults(test=True)
+args = parser.parse_args()
 
-    log_root = f"demos/{name}"
-    log_dir = "logs_train"
-    config_path = [f"demos/{name}/{name}.yaml"]
-    if test:
-        config_path.append("config/test/demo_unpaired_grouped.yaml")
+print(
+    "The training can also be launched using the following command."
+    "deepreg_train --gpu '0' "
+    f"--config_path demos/{name}/{name}.yaml "
+    f"--log_root demos/{name} "
+    "--log_dir logs_train"
+)
 
-    train(
-        gpu="0",
-        config_path=config_path,
-        gpu_allow_growth=False,
-        ckpt_path="",
-        log_root=log_root,
-        log_dir=log_dir,
-    )
+log_root = f"demos/{name}"
+log_dir = "logs_train/" + datetime.now().strftime("%Y%m%d-%H%M%S")
+config_path = [f"demos/{name}/{name}.yaml"]
+if args.test:
+    config_path.append("config/test/demo_unpaired_grouped.yaml")
 
-
-def main(args=None):
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--test",
-        help="Execute the script for test purpose",
-        dest="test",
-        action="store_true",
-    )
-    parser.add_argument(
-        "--no-test",
-        help="Execute the script for non-test purpose",
-        dest="test",
-        action="store_false",
-    )
-    parser.set_defaults(test=True)
-    args = parser.parse_args(args)
-
-    launch_training(test=args.test)
-
-
-if __name__ == "__main__":
-    main()
+train(
+    gpu="0",
+    config_path=config_path,
+    gpu_allow_growth=False,
+    ckpt_path="",
+    log_root=log_root,
+    log_dir=log_dir,
+)
