+++ demos/classical_ct_headneck_affine/demo_register.py
@@ -3,6 +3,7 @@ A DeepReg Demo for classical affine iterative pairwise registration algorithms
 """
 import argparse
 import os
+import shutil
 
 import h5py
 import tensorflow as tf
@@ -11,152 +12,144 @@ import deepreg.model.layer_util as layer_util
 import deepreg.model.loss.image as image_loss
 import deepreg.util as util
 
+parser = argparse.ArgumentParser()
+parser.add_argument(
+    "--test",
+    help="Execute the script for test purpose",
+    dest="test",
+    action="store_true",
+)
+parser.add_argument(
+    "--no-test",
+    help="Execute the script for non-test purpose",
+    dest="test",
+    action="store_false",
+)
+parser.set_defaults(test=True)
+args = parser.parse_args()
 
-def launch_registration(test: bool):
-    MAIN_PATH = os.getcwd()
-    PROJECT_DIR = "demos/classical_ct_headneck_affine"
-    os.chdir(PROJECT_DIR)
+MAIN_PATH = os.getcwd()
+PROJECT_DIR = "demos/classical_ct_headneck_affine"
+os.chdir(PROJECT_DIR)
 
-    DATA_PATH = "dataset"
-    FILE_PATH = os.path.join(DATA_PATH, "demo.h5")
+DATA_PATH = "dataset"
+FILE_PATH = os.path.join(DATA_PATH, "demo.h5")
 
-    ## registration parameters
-    image_loss_name = "ssd"
-    learning_rate = 0.01
-    total_iter = int(10) if test else int(1000)
+## registration parameters
+image_loss_name = "ssd"
+learning_rate = 0.01
+total_iter = int(10) if args.test else int(1000)
 
-    ## load image
-    if not os.path.exists(DATA_PATH):
-        raise ("Download the data using demo_data.py script")
-    if not os.path.exists(FILE_PATH):
-        raise ("Download the data using demo_data.py script")
+## load image
+if not os.path.exists(DATA_PATH):
+    raise ("Download the data using demo_data.py script")
+if not os.path.exists(FILE_PATH):
+    raise ("Download the data using demo_data.py script")
 
-    fid = h5py.File(FILE_PATH, "r")
-    fixed_image = tf.cast(tf.expand_dims(fid["image"], axis=0), dtype=tf.float32)
-    fixed_image = (fixed_image - tf.reduce_min(fixed_image)) / (
-        tf.reduce_max(fixed_image) - tf.reduce_min(fixed_image)
-    )  # normalisation to [0,1]
+fid = h5py.File(FILE_PATH, "r")
+fixed_image = tf.cast(tf.expand_dims(fid["image"], axis=0), dtype=tf.float32)
+fixed_image = (fixed_image - tf.reduce_min(fixed_image)) / (
+    tf.reduce_max(fixed_image) - tf.reduce_min(fixed_image)
+)  # normalisation to [0,1]
 
-    # generate a radomly-affine-transformed moving image
-    fixed_image_size = fixed_image.shape
-    transform_random = layer_util.random_transform_generator(batch_size=1, scale=0.2)
-    grid_ref = layer_util.get_reference_grid(grid_size=fixed_image_size[1:4])
-    grid_random = layer_util.warp_grid(grid_ref, transform_random)
-    moving_image = layer_util.resample(vol=fixed_image, loc=grid_random)
-    # warp the labels to get ground-truth using the same random affine, for validation
-    fixed_labels = tf.cast(tf.expand_dims(fid["label"], axis=0), dtype=tf.float32)
-    moving_labels = tf.stack(
-        [
-            layer_util.resample(vol=fixed_labels[..., idx], loc=grid_random)
-            for idx in range(fixed_labels.shape[4])
-        ],
-        axis=4,
-    )
+# generate a radomly-affine-transformed moving image
[…]

[…]
