### 2020-10-27 9106a80
Issue #344: add comments for parser and change classical demos to non-test by default
