+++ demos/paired_mrus_prostate/demo_train.py
@@ -6,6 +6,10 @@ from deepreg.train import train
 name = "paired_mrus_prostate"
 
 
+# parser is used to simplify testing
+# please run the script with --no-test flag to ensure non-testing mode
+# for instance:
+# python script.py --no-test
 parser = argparse.ArgumentParser()
 parser.add_argument(
     "--test",
