+++ demos/classical_mr_prostate_nonrigid/demo_register.py
@@ -13,6 +13,10 @@ import deepreg.model.loss.deform as deform_loss
 import deepreg.model.loss.image as image_loss
 import deepreg.util as util
 
+# parser is used to simplify testing
+# please run the script with --no-test flag to ensure non-testing mode
+# for instance:
+# python script.py --no-test
 parser = argparse.ArgumentParser()
 parser.add_argument(
     "--test",
@@ -26,7 +30,7 @@ parser.add_argument(
     dest="test",
     action="store_false",
 )
-parser.set_defaults(test=True)
+parser.set_defaults(test=False)
 args = parser.parse_args()
 
 MAIN_PATH = os.getcwd()
