+++ demos/paired_ct_lung/demo_train.py
@@ -5,6 +5,10 @@ from deepreg.train import train
 
 name = "paired_ct_lung"
 
+# parser is used to simplify testing
+# please run the script with --no-test flag to ensure non-testing mode
+# for instance:
+# python script.py --no-test
 parser = argparse.ArgumentParser()
 parser.add_argument(
     "--test",
