### 2020-10-27 dc239ec
Issue #344: make CLI instruction print more significant
