### 2020-10-27 bf3af4c
Issue #471: update installation instruction to resolve comments
