### 2020-10-27 84cf4e5
Issue #480: update joss paper regarding reviews
