### 2020-10-27 8c38f7f
Issue #343 minot text updates including template
