+++ demos/paired_mrus_brain/demo_predict.py
@@ -6,7 +6,7 @@ name = "paired_mrus_brain"
 
 print(
     "\n\n\n\n\n"
-    "========================================================="
+    "=========================================================\n"
     "The prediction can also be launched using the following command.\n"
     "deepreg_predict --gpu '' "
     f"--config_path demos/{name}/{name}.yaml "
