### 2020-10-27 2c34681
Issue #344 minot text updates including template
