+++ demos/grouped_mask_prostate_longitudinal/demo_predict.py
@@ -6,7 +6,7 @@ name = "grouped_mask_prostate_longitudinal"
 
 print(
     "\n\n\n\n\n"
-    "========================================================="
+    "=========================================================\n"
     "The prediction can also be launched using the following command.\n"
     "deepreg_predict --gpu '' "
     f"--config_path demos/{name}/{name}.yaml "
