+++ demos/unpaired_mr_brain/demo_predict.py
@@ -7,7 +7,7 @@ name = "unpaired_mr_brain"
 
 print(
     "\n\n\n\n\n"
-    "========================================================="
+    "=========================================================\n"
     "The prediction can also be launched using the following command.\n"
     "deepreg_predict --gpu '' "
     f"--config_path demos/{name}/{name}.yaml "
