+++ demos/paired_mrus_brain/README.md
@@ -28,7 +28,7 @@ change the current directory to the root directory of DeepReg project, i.e. `Dee
 
 ### Download data
 
-Please execute the following command to download and pre-process the data and
+Please execute the following command to download/pre-process the data and download the
 pre-trained model. By default, the downloaded data is only a partial of the original
 one. However the access to the original data is temporarily unavailable.
 
@@ -57,10 +57,10 @@ Note: The number of epochs and reduced dataset size for training will result in 
 in test accuracy so please train with the full dataset and for a greater number of
 epochs for improved results.
 
-### Launch prediction
+### Predict
 
-Please execute the following command to launch the prediction with pre-trained model.
-The prediction logs and visualization results will be saved under
+Please execute the following command to run the prediction with pre-trained model. The
+prediction logs and visualization results will be saved under
 `demos/paired_mrus_brain/logs_predict`. Check the [CLI documentation](../docs/cli.html)
 for more details about prediction output.
 
