+++ docs/source/demo/readme_template.md
@@ -27,7 +27,7 @@ change the current directory to the root directory of DeepReg project, i.e. `Dee
 
 ### Download data
 
-Please execute the following command to download and pre-process the data and
+Please execute the following command to download/pre-process the data and download the
 pre-trained model.
 
 ```bash
@@ -51,12 +51,12 @@ configuration, such as
 python demos/name/demo_train.py --no-test
 ```
 
-### Launch prediction
+### Predict
 
-Please execute the following command to launch the prediction with pre-trained model.
-The prediction logs and visualization results will be saved under
-`demos/name/logs_predict`. Check the [CLI documentation](../docs/cli.html) for more
-details about prediction output.
+Please execute the following command to run the prediction with pre-trained model. The
+prediction logs and visualization results will be saved under `demos/name/logs_predict`.
+Check the [CLI documentation](../docs/cli.html) for more details about prediction
+output.
 
 ```bash
 python demos/name/demo_predict.py
