+++ demos/unpaired_us_prostate_cv/README.md
@@ -32,7 +32,7 @@ change the current directory to the root directory of DeepReg project, i.e. `Dee
 
 ### Download data
 
-Please execute the following command to download and pre-process the data and
+Please execute the following command to download/pre-process the data and download the
 pre-trained model. Data are split into 10 folds for cross-validation.
 
 ```bash
@@ -57,10 +57,10 @@ configuration, such as
 python demos/unpaired_us_prostate_cv/demo_train.py --no-test
 ```
 
-### Launch prediction
+### Predict
 
-Please execute the following command to launch the prediction with pre-trained model.
-The prediction logs and visualization results will be saved under
+Please execute the following command to run the prediction with pre-trained model. The
+prediction logs and visualization results will be saved under
 `demos/unpaired_us_prostate_cv/logs_predict`. Check the
 [CLI documentation](../docs/cli.html) for more details about prediction output.
 
