### 2020-10-27 caa9f99
Issue #471 minor text edits and added cd command
