### 2020-10-28 DeepRegNet/DeepReg#480 @mathpluscode
@zhangfanmark Hi, we've updated the paper regarding your comments, ref https://github.com/DeepRegNet/DeepReg/commit/38caa466fee6b149c36353953e5ee262f03e576b. Please feel free to reopen this issue in case of further comments/reviews.
