### 2020-10-28 DeepRegNet/DeepReg#478 @mathpluscode
Hi @zhangfanmark , the mentioned demo has been fixed in #470. Now all demos have been tested. Please feel free to try again and let us know in case of any problems?
