+++ requirements.txt
@@ -7,7 +7,7 @@ m2r2
 matplotlib
 nibabel
 notebook
-numpy==1.18.5
+numpy==1.19.3
 pandas
 pre-commit
 pytest
