### 2020-10-30 f32764e
Issue #489: test demos depending on job type and commit msg test-demo
