### 2020-10-31 1371f88
Issue #482: alter test to cover diff batch_size case
