### 2020-10-31 9eea940
Issue #482: update changelog regarding prediction fix
