+++ CHANGELOG.md
@@ -35,6 +35,7 @@ straightforward as possible.
 
 ### Fixed
 
+- Fixed prediction error when number of samples cannot be divided by batch size exactly.
 - Fixed division by zero handling in multiple image/label losses.
 - Fixed tensor comparison in unit tests and impacted tests.
 - Removed normalization of DDF/DVF when saving in Nifti formats.
