### 2020-10-31 c745a98
Issue #482: fix typo in ci
