### 2020-10-31 0119dc8
Issue #492: update h5py version requirement
