### 2020-10-31 7f06cb3
Issue #481: check dataset is not empty for h5 and nifti
Need to add tests and refactor them
