+++ deepreg/dataset/loader/nifti_loader.py
@@ -70,6 +70,11 @@ class NiftiFileLoader(FileLoader):
                 data_path_splits += [
                     (dir_path, file_path, suffix) for file_path, suffix in data_paths
                 ]
+        if len(data_path_splits) == 0:
+            raise ValueError(
+                f"No data collected from {self.dir_paths} in NiftiFileLoader, "
+                f"please verify the path is correct."
+            )
         self.data_path_splits = sorted(data_path_splits)
 
     def set_group_structure(self):
