### 2020-10-31 3583be5
Issue #492: update h5py version requirement the same as tf
