### 2020-10-31 2770700
Issue #489: update changelog test-demo
