### 2020-11-01 fe52cc5
Issue #489: add test mode for predict but false by default
