+++ test/demo/test_demos.py
@@ -85,7 +85,7 @@ class TestDemo:
         check_files(name)
 
         # execute train, predict sequentially
-        cmds = [f"python demos/{name}/demo_{x}.py" for x in ["train", "predict"]]
+        cmds = [f"python demos/{name}/demo_{x}.py --test" for x in ["train", "predict"]]
 
         execute_commands(cmds)
 
@@ -105,7 +105,7 @@ class TestDemo:
         cmds = []
         for method in ["comb", "unsup", "weakly"]:
             cmds += [
-                f"python demos/{name}/demo_{x}.py --method {method}"
+                f"python demos/{name}/demo_{x}.py --method {method} --test"
                 for x in ["train", "predict"]
             ]
 
@@ -129,6 +129,6 @@ class TestDemo:
         check_files(name)
 
         # execute data, register
-        cmds = [f"python demos/{name}/demo_register.py"]
+        cmds = [f"python demos/{name}/demo_register.py --test"]
 
         execute_commands(cmds)
