+++ demos/paired_mrus_brain/demo_predict.py
@@ -1,8 +1,29 @@
+import argparse
 from datetime import datetime
 
 from deepreg.predict import predict
 
 name = "paired_mrus_brain"
+
+# parser is used to simplify testing, by default it is not used
+# please run the script with --no-test flag to ensure non-testing mode
+# for instance:
+# python script.py --no-test
+parser = argparse.ArgumentParser()
+parser.add_argument(
+    "--test",
+    help="Execute the script for test purpose",
+    dest="test",
+    action="store_true",
+)
+parser.add_argument(
+    "--no-test",
+    help="Execute the script for non-test purpose",
+    dest="test",
+    action="store_false",
+)
+parser.set_defaults(test=False)
+args = parser.parse_args()
 
 print(
     "\n\n\n\n\n"
@@ -21,7 +42,9 @@ print(
 log_root = f"demos/{name}"
 log_dir = "logs_predict/" + datetime.now().strftime("%Y%m%d-%H%M%S")
 ckpt_path = f"{log_root}/dataset/pretrained/learn2reg_t1_paired_train_logs/save/weights-epoch800.ckpt"
-config_path = f"{log_root}/{name}.yaml"
+config_path = [f"{log_root}/{name}.yaml"]
+if args.test:
+    config_path.append("config/test/demo_paired.yaml")
 
 predict(
     gpu="0",
