+++ .travis.yml
@@ -54,9 +54,15 @@ jobs:
     - name: "test_single_config_demo[unpaired_us_prostate_cv]"
       if: type = cron OR commit_message =~ /test-demo/
       script: pytest test/demo/ -k test_single_config_demo[unpaired_us_prostate_cv]
-    - name: "test_unpaired_ct_abdomen"
+    - name: "test_unpaired_ct_abdomen[comb]"
       if: type = cron OR commit_message =~ /test-demo/
-      script: pytest test/demo/ -k test_unpaired_ct_abdomen
+      script: pytest test/demo/ -k test_unpaired_ct_abdomen[comb]
+    - name: "test_unpaired_ct_abdomen[unsup]"
+      if: type = cron OR commit_message =~ /test-demo/
+      script: pytest test/demo/ -k test_unpaired_ct_abdomen[unsup]
+    - name: "test_unpaired_ct_abdomen[weakly]"
+      if: type = cron OR commit_message =~ /test-demo/
+      script: pytest test/demo/ -k test_unpaired_ct_abdomen[weakly]
     - name: "test_classical_demo[classical_ct_headneck_affine]"
       if: type = cron OR commit_message =~ /test-demo/
       script: pytest test/demo/ -k test_classical_demo[classical_ct_headneck_affine]
