### 2020-11-01 95ba1aa
Issue #481: refactor h5 loader tests
