+++ test/unit/test_h5_loader.py
@@ -15,333 +15,347 @@ def get_loader_h5_file_names(loader: H5FileLoader) -> List[str]:
     return [f.filename for f in loader.h5_files.values()]
 
 
-def test_init():
-    """
-    check init of H5FileLoader
-    this includes some test on set_data_structure/set_group_structure
-    as these functions are called in init
-    """
+class TestH5FileLoader:
+    def test_init_paired(self):
+        # test h5_files and data_path_splits
+        dir_paths = ["./data/test/h5/paired/test"]
+        name = "fixed_images"
+        loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=False)
+        got = [get_loader_h5_file_names(loader), loader.data_path_splits]
+        expected = [
+            ["./data/test/h5/paired/test/fixed_images.h5"],
+            [("./data/test/h5/paired/test", "case000025.nii.gz")],
+        ]
+        loader.close()
+        assert got == expected
 
-    # paired
-    # test h5_files and data_path_splits
-    dir_paths = ["./data/test/h5/paired/test"]
-    name = "fixed_images"
-    loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=False)
-    got = [get_loader_h5_file_names(loader), loader.data_path_splits]
-    expected = [
-        ["./data/test/h5/paired/test/fixed_images.h5"],
-        [("./data/test/h5/paired/test", "case000025.nii.gz")],
-    ]
-    loader.close()
-    assert got == expected
+    def test_init_unpaired(self):
+        # test h5_files and data_path_splits
+        dir_paths = ["./data/test/h5/unpaired/test"]
+        name = "images"
+        loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=False)
+        got = [get_loader_h5_file_names(loader), loader.data_path_splits]
+        expected = [
+            ["./data/test/h5/unpaired/test/images.h5"],
+            [
+                ("./data/test/h5/unpaired/test", "case000025.nii.gz"),
+                ("./data/test/h5/unpaired/test", "case000026.nii.gz"),
+            ],
+        ]
+        loader.close()
+        assert got == expected
 
-    # unpaired
-    # test h5_files and data_path_splits
-    dir_paths = ["./data/test/h5/unpaired/test"]
-    name = "images"
-    loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=False)
-    got = [get_loader_h5_file_names(loader), loader.data_path_splits]
-    expected = [
-        ["./data/test/h5/unpaired/test/images.h5"],
+    def test_init_grouped(self):
+        # test h5_files, data_path_splits and group_struct
+        dir_paths = ["./data/test/h5/grouped/test"]
+        name = "images"
+        loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=True)
+        got = [
+            get_loader_h5_file_names(loader),
+            loader.data_path_splits,
+            loader.group_struct,
+        ]
+        expected = [
+            ["./data/test/h5/grouped/test/images.h5"],
+            [
+                ("./data/test/h5/grouped/test", "1", "1"),
+                ("./data/test/h5/grouped/test", "1", "2"),
+            ],
+            [[0, 1]],
+        ]
+        loader.close()
+        assert got == expected
+
+    def test_init_multi_dirs_grouped(self):
+        # test h5_files, data_path_splits and group_struct
+        dir_paths = ["./data/test/h5/grouped/train", "./data/test/h5/grouped/test"]
+        name = "images"
+        loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=True)
+        got = [
+            get_loader_h5_file_names(loader),
+            loader.data_path_splits,
+            loader.group_struct,
+        ]
[…]
