### 2020-11-02 91a60aa
Issue #481: refactor h5 loader test
