+++ test/unit/test_h5_loader.py
@@ -1,9 +1,11 @@
 """
 Tests functionality of the H5FileLoader
 """
+import os
 from test.unit.util import is_equal_np
 from typing import List
 
+import h5py
 import numpy as np
 import pytest
 
@@ -15,347 +17,247 @@ def get_loader_h5_file_names(loader: H5FileLoader) -> List[str]:
     return [f.filename for f in loader.h5_files.values()]
 
 
-class TestH5FileLoader:
-    def test_init_paired(self):
-        # test h5_files and data_path_splits
-        dir_paths = ["./data/test/h5/paired/test"]
-        name = "fixed_images"
-        loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=False)
-        got = [get_loader_h5_file_names(loader), loader.data_path_splits]
-        expected = [
-            ["./data/test/h5/paired/test/fixed_images.h5"],
-            [("./data/test/h5/paired/test", "case000025.nii.gz")],
-        ]
-        loader.close()
-        assert got == expected
-
-    def test_init_unpaired(self):
-        # test h5_files and data_path_splits
-        dir_paths = ["./data/test/h5/unpaired/test"]
-        name = "images"
-        loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=False)
-        got = [get_loader_h5_file_names(loader), loader.data_path_splits]
-        expected = [
-            ["./data/test/h5/unpaired/test/images.h5"],
-            [
-                ("./data/test/h5/unpaired/test", "case000025.nii.gz"),
-                ("./data/test/h5/unpaired/test", "case000026.nii.gz"),
-            ],
-        ]
-        loader.close()
-        assert got == expected
-
-    def test_init_grouped(self):
-        # test h5_files, data_path_splits and group_struct
-        dir_paths = ["./data/test/h5/grouped/test"]
-        name = "images"
-        loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=True)
-        got = [
-            get_loader_h5_file_names(loader),
-            loader.data_path_splits,
-            loader.group_struct,
-        ]
-        expected = [
-            ["./data/test/h5/grouped/test/images.h5"],
-            [
-                ("./data/test/h5/grouped/test", "1", "1"),
-                ("./data/test/h5/grouped/test", "1", "2"),
-            ],
-            [[0, 1]],
-        ]
-        loader.close()
-        assert got == expected
-
-    def test_init_multi_dirs_grouped(self):
-        # test h5_files, data_path_splits and group_struct
+def get_loader(loader_name):
+    if loader_name in [
+        "paired",
+        "unpaired",
+        "grouped",
+    ]:
+        dir_paths = [f"./data/test/h5/{loader_name}/test"]
+        name = "fixed_images" if loader_name == "paired" else "images"
+        grouped = loader_name == "grouped"
+    elif loader_name == "multi_dirs_grouped":
         dir_paths = ["./data/test/h5/grouped/train", "./data/test/h5/grouped/test"]
         name = "images"
-        loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=True)
+        grouped = True
+    else:
+        raise ValueError
+    loader = H5FileLoader(dir_paths=dir_paths, name=name, grouped=grouped)
+    return loader
+
+
+class TestH5FileLoader:
+    dir_paths = ["./data/test/h5/paired/test"]
+    name = "fixed_images"
+
+    @pytest.mark.parametrize(
+        "name,expected",
+        [
+            (
+                "paired",
+                [
+                    ["./data/test/h5/paired/test/fixed_images.h5"],
+                    [("./data/test/h5/paired/test", "case000025.nii.gz")],
+                    None,
+                ],
+            ),
+            (
+                "unpaired",
+                [
[…]
