+++ deepreg/dataset/loader/h5_loader.py
@@ -22,8 +22,8 @@ class H5FileLoader(FileLoader):
         self.h5_files = None
         self.data_path_splits = None
         self.set_data_structure()
+        self.group_struct = None
         if self.grouped:
-            self.group_struct = None
             self.set_group_structure()
 
     def set_data_structure(self):
