### 2020-11-02 e26c532
Issue #489: refactor nifti loader test
