+++ test/unit/test_nifti_loader.py
@@ -1,6 +1,7 @@
 """
 Tests functionality of the NiftiFileLoader
 """
+import os
 from test.unit.util import is_equal_np
 
 import numpy as np
@@ -9,331 +10,323 @@ import pytest
 from deepreg.dataset.loader.nifti_loader import NiftiFileLoader, load_nifti_file
 
 
-def test_load_nifti_file():
+def get_loader(loader_name):
+    if loader_name in [
+        "paired",
+        "unpaired",
+        "grouped",
+    ]:
+        dir_paths = [f"./data/test/nifti/{loader_name}/test"]
+        name = "fixed_images" if loader_name == "paired" else "images"
+        grouped = loader_name == "grouped"
+    elif loader_name == "multi_dirs_grouped":
+        dir_paths = [
+            "./data/test/nifti/grouped/train",
+            "./data/test/nifti/grouped/test",
+        ]
+        name = "images"
+        grouped = True
+    else:
+        raise ValueError
+    loader = NiftiFileLoader(dir_paths=dir_paths, name=name, grouped=grouped)
+    return loader
+
+
+@pytest.mark.parametrize(
+    "path,shape",
+    [
+        ("./data/test/nifti/paired/test/fixed_images/case000026.nii.gz", (44, 59, 41)),
+        ("./data/test/nifti/unit_test/case000026.nii", (44, 59, 41)),
+    ],
+)
+def test_load_nifti_file(path, shape):
     """
     check if the nifti files can be correctly loaded
     """
+    arr = load_nifti_file(file_path=path)
+    assert arr.shape == shape
 
-    # nii.gz
-    nii_gz_filepath = "./data/test/nifti/paired/test/fixed_images/case000026.nii.gz"
-    load_nifti_file(file_path=nii_gz_filepath)
 
-    # nii
-    nii_filepath = "./data/test/nifti/unit_test/case000026.nii"
-    load_nifti_file(file_path=nii_filepath)
-
-    # wrong file type
+def test_load_nifti_file_err():
     h5_filepath = "./data/test/h5/paired/test/fixed_images.h5"
     with pytest.raises(ValueError) as err_info:
         load_nifti_file(file_path=h5_filepath)
     assert "Nifti file path must end with .nii or .nii.gz" in str(err_info.value)
 
 
-def test_init():
-    """
-    check init of NiftiFileLoader
-    this includes some test on set_data_structure/set_group_structure
-    as these functions are called in init
-    """
-
-    # paired
-    # test data_path_splits
-    dir_paths = ["./data/test/nifti/paired/test"]
-    name = "fixed_images"
-    loader = NiftiFileLoader(dir_paths=dir_paths, name=name, grouped=False)
-    got = loader.data_path_splits
-    expected = [
-        ("./data/test/nifti/paired/test", "case000025", "nii.gz"),
-        ("./data/test/nifti/paired/test", "case000026", "nii.gz"),
-    ]
-    loader.close()
-    assert got == expected
-
-    # unpaired
-    # test data_path_splits
-    dir_paths = ["./data/test/nifti/unpaired/test"]
-    name = "images"
-    loader = NiftiFileLoader(dir_paths=dir_paths, name=name, grouped=False)
-    got = loader.data_path_splits
-    expected = [
-        ("./data/test/nifti/unpaired/test", "case000025", "nii.gz"),
-        ("./data/test/nifti/unpaired/test", "case000026", "nii"),
-    ]
-    loader.close()
-    assert got == expected
-
-    # grouped
-    # test data_path_splits and group_struct
-    dir_paths = ["./data/test/nifti/grouped/test"]
-    name = "images"
-    loader = NiftiFileLoader(dir_paths=dir_paths, name=name, grouped=True)
[…]
