### 2020-11-02 0ec6b5e
Issue #489: update changelog
