### 2020-11-02 1f1e899
Update h5py requirement from <2.11.0,>=2.10.0 to >=2.10.0,<3.1.0
Updates the requirements on [h5py](https://github.com/h5py/h5py) to permit the latest version.
- [Release notes](https://github.com/h5py/h5py/releases)
- [Changelog](https://github.com/h5py/h5py/blob/master/docs/release_guide.rst)
- [Commits](https://github.com/h5py/h5py/compare/2.10.0...3.0.0)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
