### 2020-11-02 DeepRegNet/DeepReg#479 @mathpluscode
Hi @zhangfanmark  thank you again for your review. I'm closing this as we are preparing for the acceptance and the release of `v0.1.0`.
