### 2020-11-02 DeepRegNet/DeepReg#499 @mathpluscode
Release v0.1.0
## Subject of the issue

As mentioned in https://github.com/openjournals/joss-reviews/issues/2705, for preparing the acceptance to JOSS, we will release the version `v0.1.0` therefore some text/version/docs need to be updated.
