### 2020-11-02 a3c5288
Issue #499: update docs regarding release v0.1.0
