+++ CHANGELOG.md
@@ -1,14 +1,12 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
+All notable changes to this project will be documented in this file. It's a team effort
+to make them as straightforward as possible.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project
 adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
-
-Here we write upgrading notes for brands. It's a team effort to make them as
-straightforward as possible.
+## [0.1.0] - 2020-11-03
 
 ### Added
 
