### 2020-11-02 DeepRegNet/DeepReg#499 @mathpluscode
Reopen as need to modify Zenodo metadata
