### 2020-11-02 7576329
Issue #499: add zenodo config
