+++ README.md
@@ -10,6 +10,7 @@
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![codecov](https://codecov.io/gh/DeepRegNet/DeepReg/branch/main/graph/badge.svg)](https://codecov.io/gh/DeepRegNet/DeepReg)
 [![status](https://joss.theoj.org/papers/7e6de472bc82a70d7618e23f618960b3/status.svg)](https://joss.theoj.org/papers/7e6de472bc82a70d7618e23f618960b3)
+[![DOI](https://zenodo.org/badge/269365590.svg)](https://zenodo.org/badge/latestdoi/269365590)
 
 **DeepReg is a freely available, community-supported open-source toolkit for research
 and education in medical image registration using deep learning.**
