### 2020-11-02 04fb496
Issue #499: update description and license
