### 2020-11-02 #2705 @mathpluscode
> @mathpluscode this looks like it is ready to process for acceptance. At this point can you do the following:
>
> * [ ]  Can you archive a copy of the reviewed software on Zenodo and report back here with the DOI? (some find this approach useful: https://guides.github.com/activities/citable-code/).
> * [ ]  Please check that the Zenodo archive metadata like the **title**, **authors**, and **version** matches that of the paper. You may need to ammend/edit this information manually.
> * [ ]  Can you confirm if `v0.1.0` is the most appropriate version tag for this work, or has the version moved on?
>
> Thanks

Hi @Kevin-Mattheus-Moerman,
- Here is the archive of DeepReg `v0.1.0` [![DOI](https://zenodo.org/badge/269365590.svg)](https://zenodo.org/badge/latestdoi/269365590)
- Yes we've corrected the title, authors and version
- Yes we confirm `v0.1.0` is the correct version tag.
