+++ requirements.txt
@@ -16,7 +16,7 @@ pytest-dependency
 pyyaml>=5.3.1
 scipy
 seed-isort-config
-sphinx==3.2.1
+sphinx==3.3.0
 sphinx-notfound-page
 sphinx-rtd-theme
 sphinx-tabs
