### 2020-11-03 0b36293
Bump numpy from 1.18.5 to 1.19.4
Bumps [numpy](https://github.com/numpy/numpy) from 1.18.5 to 1.19.4.
- [Release notes](https://github.com/numpy/numpy/releases)
- [Changelog](https://github.com/numpy/numpy/blob/master/doc/HOWTO_RELEASE.rst.txt)
- [Commits](https://github.com/numpy/numpy/compare/v1.18.5...v1.19.4)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
