### 2020-11-03 34f8c50
Issue #505: Updated to correct stable tag for PyPI releases
