+++ setup.py
@@ -36,7 +36,7 @@ setup(
         ]
     },
     classifiers=[
-        "Development Status :: 5 - Stable",
+        "Development Status :: 5 - Production/Stable",
         "Environment :: Console",
         "Environment :: GPU",
         "Environment :: MacOS X",
